Make an MultiFactor

Construct an MultiFactor object from a validly shaped data frame or list of such data.frames.

Helper function that takes an MultiFactor and returns a sparse biadjacency Matrix with link df names as rownames and id names as colnames. Called internally.

following S4Vectors:levels.Rle()

Usage

MultiFactor(x)

asMultiFactor(x)

linkMatrix(x)

levels.MultiFactor(x)

Arguments

x

any object, most likely list of data frames.

Value

a sparse biadjacency Matrix with link df names as rownames and id names as colnames

See also

Examples

MultiFactor( kegg_link( ) )
#> A list of class MultiFactor,
#>     3 feature types across 2 edge lists.
#> 
#>          ec   ko  cpd
#> ec2ko  7802 5934    .
#> ec2cpd    . 5581 8817
#> 
#> Values represent unique feature names in that edge list.