Methods for pairwise operations on AnansiWeb objects

Methods to run pairwise analysis on multi-modal data.

Arguments

X, x

input, AnansiWeb object

...

additional arguments

FUN

a function with at least two arguments. The variables x and y, in order, refer to the corresponding values of feature pairs in tableX and tableY.

MoreArgs, SIMPLIFY, USE.NAMES

see ?base::mapply

which

integer matrix, indicating pair positions in x@tableY and x@tableX, respectively. If NULL (default): Matrix::which(x@dictionary, TRUE).

with.metadata

Logical scalar whether to append metadata to output

Value

pairs: A two-column array index, corresponding to i,j coordinates in matrix notation.
getFeaturePairs: A list of data.frames with the paired data

Examples

web <- randomWeb()
# Extract data.frames in pairs (only show first)
getFeaturePairs(web)[1L]
#> [[1]]
#>                              y_1        x_1
#> anansi_ID_sample_1_1  -1.5727175  0.9347076
#> anansi_ID_sample_2_1  -0.9864274 -0.4726698
#> anansi_ID_sample_3_1   1.9835123  1.5830132
#> anansi_ID_sample_4_1  -1.8513784  0.5977542
#> anansi_ID_sample_5_1  -0.9097882  0.4660846
#> anansi_ID_sample_6_1  -1.9514400 -0.7443675
#> anansi_ID_sample_7_1  -0.8002765  0.9590553
#> anansi_ID_sample_8_1  -1.8693616  0.1526371
#> anansi_ID_sample_9_1  -0.7505715 -0.7693687
#> anansi_ID_sample_10_1 -0.5909384  1.3930807
#> 

pairs(x = web)
#>       [,1] [,2]
#>  [1,]    1    1
#>  [2,]    3    1
#>  [3,]    9    1
#>  [4,]   11    1
#>  [5,]   12    1
#>  [6,]    1    2
#>  [7,]    2    2
#>  [8,]    5    2
#>  [9,]    7    2
#> [10,]    8    2
#> [11,]    9    2
#> [12,]   11    2
#> [13,]    2    3
#> [14,]    3    3
#> [15,]    4    3
#> [16,]    7    3
#> [17,]    8    3
#> [18,]   10    3
#> [19,]   12    3
#> [20,]    1    4
#> [21,]    2    4
#> [22,]    3    4
#> [23,]    6    4
#> [24,]    8    4
#> [25,]    9    4
#> [26,]   12    4
#> [27,]    6    5
#> [28,]    9    5
#> [29,]   10    5
#> [30,]   12    5
#> [31,]    1    6
#> [32,]    4    6
#> [33,]    5    6
#> [34,]    8    6
#> [35,]   10    6
#> [36,]   11    6
#> [37,]    2    7
#> [38,]    3    7
#> [39,]    4    7
#> [40,]    6    7
#> [41,]   11    7
#> [42,]   12    7
#> [43,]    1    8
#> [44,]    4    8
#> [45,]    5    8
#> [46,]    8    8
#> [47,]    9    8
#> [48,]   11    8

getFeaturePairs(x = web, which = NULL, with.metadata = FALSE)
#> [[1]]
#>                              y_1        x_1
#> anansi_ID_sample_1_1  -1.5727175  0.9347076
#> anansi_ID_sample_2_1  -0.9864274 -0.4726698
#> anansi_ID_sample_3_1   1.9835123  1.5830132
#> anansi_ID_sample_4_1  -1.8513784  0.5977542
#> anansi_ID_sample_5_1  -0.9097882  0.4660846
#> anansi_ID_sample_6_1  -1.9514400 -0.7443675
#> anansi_ID_sample_7_1  -0.8002765  0.9590553
#> anansi_ID_sample_8_1  -1.8693616  0.1526371
#> anansi_ID_sample_9_1  -0.7505715 -0.7693687
#> anansi_ID_sample_10_1 -0.5909384  1.3930807
#> 
#> [[2]]
#>                              y_3        x_1
#> anansi_ID_sample_1_1  -0.6127534  0.9347076
#> anansi_ID_sample_2_1   0.1606625 -0.4726698
#> anansi_ID_sample_3_1  -0.6623596  1.5830132
#> anansi_ID_sample_4_1  -0.3348517  0.5977542
#> anansi_ID_sample_5_1   0.6230116  0.4660846
#> anansi_ID_sample_6_1   1.0281943 -0.7443675
#> anansi_ID_sample_7_1  -1.1345782  0.9590553
#> anansi_ID_sample_8_1   0.9169111  0.1526371
#> anansi_ID_sample_9_1   1.2138748 -0.7693687
#> anansi_ID_sample_10_1 -0.6859325  1.3930807
#> 
#> [[3]]
#>                               y_9        x_1
#> anansi_ID_sample_1_1   0.35846241  0.9347076
#> anansi_ID_sample_2_1   0.04818886 -0.4726698
#> anansi_ID_sample_3_1   1.13017228  1.5830132
#> anansi_ID_sample_4_1  -0.52763231  0.5977542
#> anansi_ID_sample_5_1   0.47957533  0.4660846
#> anansi_ID_sample_6_1   1.41752930 -0.7443675
#> anansi_ID_sample_7_1  -0.08009077  0.9590553
#> anansi_ID_sample_8_1  -0.39418700  0.1526371
#> anansi_ID_sample_9_1   0.52525764 -0.7693687
#> anansi_ID_sample_10_1 -0.79593510  1.3930807
#> 
#> [[4]]
#>                             y_11        x_1
#> anansi_ID_sample_1_1  -1.6634113  0.9347076
#> anansi_ID_sample_2_1   1.9156908 -0.4726698
#> anansi_ID_sample_3_1  -0.8168944  1.5830132
#> anansi_ID_sample_4_1   0.3836505  0.5977542
#> anansi_ID_sample_5_1  -0.4586323  0.4660846
#> anansi_ID_sample_6_1  -0.7157040 -0.7443675
#> anansi_ID_sample_7_1   0.4295485  0.9590553
#> anansi_ID_sample_8_1  -1.1698527  0.1526371
#> anansi_ID_sample_9_1   1.3220883 -0.7693687
#> anansi_ID_sample_10_1  1.3659646  1.3930807
#> 
#> [[5]]
#>                              y_12        x_1
#> anansi_ID_sample_1_1  -0.24972371  0.9347076
#> anansi_ID_sample_2_1  -0.42472135 -0.4726698
#> anansi_ID_sample_3_1   1.09843934  1.5830132
#> anansi_ID_sample_4_1   0.67092304  0.5977542
#> anansi_ID_sample_5_1   0.03095466  0.4660846
#> anansi_ID_sample_6_1   2.36851196 -0.7443675
#> anansi_ID_sample_7_1   2.58578811  0.9590553
#> anansi_ID_sample_8_1   1.21228095  0.1526371
#> anansi_ID_sample_9_1  -1.25119684 -0.7693687
#> anansi_ID_sample_10_1  0.40026604  1.3930807
#> 
#> [[6]]
#>                              y_1         x_2
#> anansi_ID_sample_1_1  -1.5727175  0.91866877
#> anansi_ID_sample_2_1  -0.9864274  1.01608706
#> anansi_ID_sample_3_1   1.9835123  0.53463544
#> anansi_ID_sample_4_1  -1.8513784 -0.98762856
#> anansi_ID_sample_5_1  -0.9097882  1.18779227
#> anansi_ID_sample_6_1  -1.9514400 -0.51754221
#> anansi_ID_sample_7_1  -0.8002765 -0.25956025
#> anansi_ID_sample_8_1  -1.8693616 -0.32806467
#> anansi_ID_sample_9_1  -0.7505715  0.07343239
#> anansi_ID_sample_10_1 -0.5909384 -0.24786302
#> 
#> [[7]]
#>                               y_2         x_2
#> anansi_ID_sample_1_1  -0.74209927  0.91866877
#> anansi_ID_sample_2_1   0.69351208  1.01608706
#> anansi_ID_sample_3_1  -0.05946397  0.53463544
#> anansi_ID_sample_4_1  -1.86389510 -0.98762856
#> anansi_ID_sample_5_1  -1.27450892  1.18779227
#> anansi_ID_sample_6_1  -1.78177081 -0.51754221
#> anansi_ID_sample_7_1  -0.50859339 -0.25956025
#> anansi_ID_sample_8_1  -1.73687829 -0.32806467
#> anansi_ID_sample_9_1   0.04048130  0.07343239
#> anansi_ID_sample_10_1 -0.12417946 -0.24786302
#> 
#> [[8]]
#>                              y_5         x_2
#> anansi_ID_sample_1_1   0.6314438  0.91866877
#> anansi_ID_sample_2_1  -0.4833983  1.01608706
#> anansi_ID_sample_3_1  -1.5263339  0.53463544
#> anansi_ID_sample_4_1   0.6168423 -0.98762856
#> anansi_ID_sample_5_1   1.2292868  1.18779227
#> anansi_ID_sample_6_1   0.1574530 -0.51754221
#> anansi_ID_sample_7_1   1.4063421 -0.25956025
#> anansi_ID_sample_8_1   0.3968009 -0.32806467
#> anansi_ID_sample_9_1   0.3264230  0.07343239
#> anansi_ID_sample_10_1 -1.4593433 -0.24786302
#> 
#> [[9]]
#>                               y_7         x_2
#> anansi_ID_sample_1_1  -0.75096428  0.91866877
#> anansi_ID_sample_2_1  -1.25336629  1.01608706
#> anansi_ID_sample_3_1  -1.03623626  0.53463544
#> anansi_ID_sample_4_1  -0.02943857 -0.98762856
#> anansi_ID_sample_5_1  -0.27494378  1.18779227
#> anansi_ID_sample_6_1   0.48545147 -0.51754221
#> anansi_ID_sample_7_1   0.91698203 -0.25956025
#> anansi_ID_sample_8_1  -0.98534392 -0.32806467
#> anansi_ID_sample_9_1  -1.49118975  0.07343239
#> anansi_ID_sample_10_1  0.81959245 -0.24786302
#> 
#> [[10]]
#>                               y_8         x_2
#> anansi_ID_sample_1_1   1.01340453  0.91866877
#> anansi_ID_sample_2_1   1.05360525  1.01608706
#> anansi_ID_sample_3_1  -0.07142864  0.53463544
#> anansi_ID_sample_4_1   0.95244461 -0.98762856
#> anansi_ID_sample_5_1  -0.38403752  1.18779227
#> anansi_ID_sample_6_1  -0.41775999 -0.51754221
#> anansi_ID_sample_7_1  -0.47026817 -0.25956025
#> anansi_ID_sample_8_1  -1.69018029 -0.32806467
#> anansi_ID_sample_9_1  -0.94549766  0.07343239
#> anansi_ID_sample_10_1 -0.34301196 -0.24786302
#> 
#> [[11]]
#>                               y_9         x_2
#> anansi_ID_sample_1_1   0.35846241  0.91866877
#> anansi_ID_sample_2_1   0.04818886  1.01608706
#> anansi_ID_sample_3_1   1.13017228  0.53463544
#> anansi_ID_sample_4_1  -0.52763231 -0.98762856
#> anansi_ID_sample_5_1   0.47957533  1.18779227
#> anansi_ID_sample_6_1   1.41752930 -0.51754221
#> anansi_ID_sample_7_1  -0.08009077 -0.25956025
#> anansi_ID_sample_8_1  -0.39418700 -0.32806467
#> anansi_ID_sample_9_1   0.52525764  0.07343239
#> anansi_ID_sample_10_1 -0.79593510 -0.24786302
#> 
#> [[12]]
#>                             y_11         x_2
#> anansi_ID_sample_1_1  -1.6634113  0.91866877
#> anansi_ID_sample_2_1   1.9156908  1.01608706
#> anansi_ID_sample_3_1  -0.8168944  0.53463544
#> anansi_ID_sample_4_1   0.3836505 -0.98762856
#> anansi_ID_sample_5_1  -0.4586323  1.18779227
#> anansi_ID_sample_6_1  -0.7157040 -0.51754221
#> anansi_ID_sample_7_1   0.4295485 -0.25956025
#> anansi_ID_sample_8_1  -1.1698527 -0.32806467
#> anansi_ID_sample_9_1   1.3220883  0.07343239
#> anansi_ID_sample_10_1  1.3659646 -0.24786302
#> 
#> [[13]]
#>                               y_2         x_3
#> anansi_ID_sample_1_1  -0.74209927 -1.37386226
#> anansi_ID_sample_2_1   0.69351208 -0.04044582
#> anansi_ID_sample_3_1  -0.05946397  0.42153824
#> anansi_ID_sample_4_1  -1.86389510  0.20159751
#> anansi_ID_sample_5_1  -1.27450892 -1.69719192
#> anansi_ID_sample_6_1  -1.78177081  0.64228768
#> anansi_ID_sample_7_1  -0.50859339 -0.99523961
#> anansi_ID_sample_8_1  -1.73687829  0.96381390
#> anansi_ID_sample_9_1   0.04048130 -1.65603723
#> anansi_ID_sample_10_1 -0.12417946  1.07086109
#> 
#> [[14]]
#>                              y_3         x_3
#> anansi_ID_sample_1_1  -0.6127534 -1.37386226
#> anansi_ID_sample_2_1   0.1606625 -0.04044582
#> anansi_ID_sample_3_1  -0.6623596  0.42153824
#> anansi_ID_sample_4_1  -0.3348517  0.20159751
#> anansi_ID_sample_5_1   0.6230116 -1.69719192
#> anansi_ID_sample_6_1   1.0281943  0.64228768
#> anansi_ID_sample_7_1  -1.1345782 -0.99523961
#> anansi_ID_sample_8_1   0.9169111  0.96381390
#> anansi_ID_sample_9_1   1.2138748 -1.65603723
#> anansi_ID_sample_10_1 -0.6859325  1.07086109
#> 
#> [[15]]
#>                              y_4         x_3
#> anansi_ID_sample_1_1  -1.0945682 -1.37386226
#> anansi_ID_sample_2_1   0.3648741 -0.04044582
#> anansi_ID_sample_3_1  -1.0080171  0.42153824
#> anansi_ID_sample_4_1   0.5549964  0.20159751
#> anansi_ID_sample_5_1   0.4360659 -1.69719192
#> anansi_ID_sample_6_1   0.1053733  0.64228768
#> anansi_ID_sample_7_1  -0.2537379 -0.99523961
#> anansi_ID_sample_8_1  -0.9764912  0.96381390
#> anansi_ID_sample_9_1   0.3942178 -1.65603723
#> anansi_ID_sample_10_1  1.0149030  1.07086109
#> 
#> [[16]]
#>                               y_7         x_3
#> anansi_ID_sample_1_1  -0.75096428 -1.37386226
#> anansi_ID_sample_2_1  -1.25336629 -0.04044582
#> anansi_ID_sample_3_1  -1.03623626  0.42153824
#> anansi_ID_sample_4_1  -0.02943857  0.20159751
#> anansi_ID_sample_5_1  -0.27494378 -1.69719192
#> anansi_ID_sample_6_1   0.48545147  0.64228768
#> anansi_ID_sample_7_1   0.91698203 -0.99523961
#> anansi_ID_sample_8_1  -0.98534392  0.96381390
#> anansi_ID_sample_9_1  -1.49118975 -1.65603723
#> anansi_ID_sample_10_1  0.81959245  1.07086109
#> 
#> [[17]]
#>                               y_8         x_3
#> anansi_ID_sample_1_1   1.01340453 -1.37386226
#> anansi_ID_sample_2_1   1.05360525 -0.04044582
#> anansi_ID_sample_3_1  -0.07142864  0.42153824
#> anansi_ID_sample_4_1   0.95244461  0.20159751
#> anansi_ID_sample_5_1  -0.38403752 -1.69719192
#> anansi_ID_sample_6_1  -0.41775999  0.64228768
#> anansi_ID_sample_7_1  -0.47026817 -0.99523961
#> anansi_ID_sample_8_1  -1.69018029  0.96381390
#> anansi_ID_sample_9_1  -0.94549766 -1.65603723
#> anansi_ID_sample_10_1 -0.34301196  1.07086109
#> 
#> [[18]]
#>                             y_10         x_3
#> anansi_ID_sample_1_1   0.2450964 -1.37386226
#> anansi_ID_sample_2_1  -0.7793198 -0.04044582
#> anansi_ID_sample_3_1   0.5949239  0.42153824
#> anansi_ID_sample_4_1   1.1088909  0.20159751
#> anansi_ID_sample_5_1  -0.9429690 -1.69719192
#> anansi_ID_sample_6_1   0.7003440  0.64228768
#> anansi_ID_sample_7_1  -0.4246748 -0.99523961
#> anansi_ID_sample_8_1  -1.1431376  0.96381390
#> anansi_ID_sample_9_1   0.2334402 -1.65603723
#> anansi_ID_sample_10_1 -0.1892366  1.07086109
#> 
#> [[19]]
#>                              y_12         x_3
#> anansi_ID_sample_1_1  -0.24972371 -1.37386226
#> anansi_ID_sample_2_1  -0.42472135 -0.04044582
#> anansi_ID_sample_3_1   1.09843934  0.42153824
#> anansi_ID_sample_4_1   0.67092304  0.20159751
#> anansi_ID_sample_5_1   0.03095466 -1.69719192
#> anansi_ID_sample_6_1   2.36851196  0.64228768
#> anansi_ID_sample_7_1   2.58578811 -0.99523961
#> anansi_ID_sample_8_1   1.21228095  0.96381390
#> anansi_ID_sample_9_1  -1.25119684 -1.65603723
#> anansi_ID_sample_10_1  0.40026604  1.07086109
#> 
#> [[20]]
#>                              y_1         x_4
#> anansi_ID_sample_1_1  -1.5727175 -0.10902636
#> anansi_ID_sample_2_1  -0.9864274  1.89918639
#> anansi_ID_sample_3_1   1.9835123 -1.13703073
#> anansi_ID_sample_4_1  -1.8513784 -0.27971976
#> anansi_ID_sample_5_1  -0.9097882 -0.89412905
#> anansi_ID_sample_6_1  -1.9514400  0.13670185
#> anansi_ID_sample_7_1  -0.8002765 -0.74916542
#> anansi_ID_sample_8_1  -1.8693616  0.51819908
#> anansi_ID_sample_9_1  -0.7505715 -0.19233721
#> anansi_ID_sample_10_1 -0.5909384  0.02880981
#> 
#> [[21]]
#>                               y_2         x_4
#> anansi_ID_sample_1_1  -0.74209927 -0.10902636
#> anansi_ID_sample_2_1   0.69351208  1.89918639
#> anansi_ID_sample_3_1  -0.05946397 -1.13703073
#> anansi_ID_sample_4_1  -1.86389510 -0.27971976
#> anansi_ID_sample_5_1  -1.27450892 -0.89412905
#> anansi_ID_sample_6_1  -1.78177081  0.13670185
#> anansi_ID_sample_7_1  -0.50859339 -0.74916542
#> anansi_ID_sample_8_1  -1.73687829  0.51819908
#> anansi_ID_sample_9_1   0.04048130 -0.19233721
#> anansi_ID_sample_10_1 -0.12417946  0.02880981
#> 
#> [[22]]
#>                              y_3         x_4
#> anansi_ID_sample_1_1  -0.6127534 -0.10902636
#> anansi_ID_sample_2_1   0.1606625  1.89918639
#> anansi_ID_sample_3_1  -0.6623596 -1.13703073
#> anansi_ID_sample_4_1  -0.3348517 -0.27971976
#> anansi_ID_sample_5_1   0.6230116 -0.89412905
#> anansi_ID_sample_6_1   1.0281943  0.13670185
#> anansi_ID_sample_7_1  -1.1345782 -0.74916542
#> anansi_ID_sample_8_1   0.9169111  0.51819908
#> anansi_ID_sample_9_1   1.2138748 -0.19233721
#> anansi_ID_sample_10_1 -0.6859325  0.02880981
#> 
#> [[23]]
#>                               y_6         x_4
#> anansi_ID_sample_1_1  -0.79929553 -0.10902636
#> anansi_ID_sample_2_1   0.47001280  1.89918639
#> anansi_ID_sample_3_1  -1.15018052 -1.13703073
#> anansi_ID_sample_4_1  -0.27159672 -0.27971976
#> anansi_ID_sample_5_1   0.45742422 -0.89412905
#> anansi_ID_sample_6_1  -0.01695794  0.13670185
#> anansi_ID_sample_7_1  -0.54159436 -0.74916542
#> anansi_ID_sample_8_1   0.87581342  0.51819908
#> anansi_ID_sample_9_1   0.74083815 -0.19233721
#> anansi_ID_sample_10_1 -0.16481239  0.02880981
#> 
#> [[24]]
#>                               y_8         x_4
#> anansi_ID_sample_1_1   1.01340453 -0.10902636
#> anansi_ID_sample_2_1   1.05360525  1.89918639
#> anansi_ID_sample_3_1  -0.07142864 -1.13703073
#> anansi_ID_sample_4_1   0.95244461 -0.27971976
#> anansi_ID_sample_5_1  -0.38403752 -0.89412905
#> anansi_ID_sample_6_1  -0.41775999  0.13670185
#> anansi_ID_sample_7_1  -0.47026817 -0.74916542
#> anansi_ID_sample_8_1  -1.69018029  0.51819908
#> anansi_ID_sample_9_1  -0.94549766 -0.19233721
#> anansi_ID_sample_10_1 -0.34301196  0.02880981
#> 
#> [[25]]
#>                               y_9         x_4
#> anansi_ID_sample_1_1   0.35846241 -0.10902636
#> anansi_ID_sample_2_1   0.04818886  1.89918639
#> anansi_ID_sample_3_1   1.13017228 -1.13703073
#> anansi_ID_sample_4_1  -0.52763231 -0.27971976
#> anansi_ID_sample_5_1   0.47957533 -0.89412905
#> anansi_ID_sample_6_1   1.41752930  0.13670185
#> anansi_ID_sample_7_1  -0.08009077 -0.74916542
#> anansi_ID_sample_8_1  -0.39418700  0.51819908
#> anansi_ID_sample_9_1   0.52525764 -0.19233721
#> anansi_ID_sample_10_1 -0.79593510  0.02880981
#> 
#> [[26]]
#>                              y_12         x_4
#> anansi_ID_sample_1_1  -0.24972371 -0.10902636
#> anansi_ID_sample_2_1  -0.42472135  1.89918639
#> anansi_ID_sample_3_1   1.09843934 -1.13703073
#> anansi_ID_sample_4_1   0.67092304 -0.27971976
#> anansi_ID_sample_5_1   0.03095466 -0.89412905
#> anansi_ID_sample_6_1   2.36851196  0.13670185
#> anansi_ID_sample_7_1   2.58578811 -0.74916542
#> anansi_ID_sample_8_1   1.21228095  0.51819908
#> anansi_ID_sample_9_1  -1.25119684 -0.19233721
#> anansi_ID_sample_10_1  0.40026604  0.02880981
#> 
#> [[27]]
#>                               y_6         x_5
#> anansi_ID_sample_1_1  -0.79929553  0.35859089
#> anansi_ID_sample_2_1   0.47001280 -0.02899503
#> anansi_ID_sample_3_1  -1.15018052  1.14704207
#> anansi_ID_sample_4_1  -0.27159672  0.37358894
#> anansi_ID_sample_5_1   0.45742422  0.32333921
#> anansi_ID_sample_6_1  -0.01695794 -0.82981932
#> anansi_ID_sample_7_1  -0.54159436  1.39446258
#> anansi_ID_sample_8_1   0.87581342 -0.19154358
#> anansi_ID_sample_9_1   0.74083815  0.27227702
#> anansi_ID_sample_10_1 -0.16481239 -1.08165901
#> 
#> [[28]]
#>                               y_9         x_5
#> anansi_ID_sample_1_1   0.35846241  0.35859089
#> anansi_ID_sample_2_1   0.04818886 -0.02899503
#> anansi_ID_sample_3_1   1.13017228  1.14704207
#> anansi_ID_sample_4_1  -0.52763231  0.37358894
#> anansi_ID_sample_5_1   0.47957533  0.32333921
#> anansi_ID_sample_6_1   1.41752930 -0.82981932
#> anansi_ID_sample_7_1  -0.08009077  1.39446258
#> anansi_ID_sample_8_1  -0.39418700 -0.19154358
#> anansi_ID_sample_9_1   0.52525764  0.27227702
#> anansi_ID_sample_10_1 -0.79593510 -1.08165901
#> 
#> [[29]]
#>                             y_10         x_5
#> anansi_ID_sample_1_1   0.2450964  0.35859089
#> anansi_ID_sample_2_1  -0.7793198 -0.02899503
#> anansi_ID_sample_3_1   0.5949239  1.14704207
#> anansi_ID_sample_4_1   1.1088909  0.37358894
#> anansi_ID_sample_5_1  -0.9429690  0.32333921
#> anansi_ID_sample_6_1   0.7003440 -0.82981932
#> anansi_ID_sample_7_1  -0.4246748  1.39446258
#> anansi_ID_sample_8_1  -1.1431376 -0.19154358
#> anansi_ID_sample_9_1   0.2334402  0.27227702
#> anansi_ID_sample_10_1 -0.1892366 -1.08165901
#> 
#> [[30]]
#>                              y_12         x_5
#> anansi_ID_sample_1_1  -0.24972371  0.35859089
#> anansi_ID_sample_2_1  -0.42472135 -0.02899503
#> anansi_ID_sample_3_1   1.09843934  1.14704207
#> anansi_ID_sample_4_1   0.67092304  0.37358894
#> anansi_ID_sample_5_1   0.03095466  0.32333921
#> anansi_ID_sample_6_1   2.36851196 -0.82981932
#> anansi_ID_sample_7_1   2.58578811  1.39446258
#> anansi_ID_sample_8_1   1.21228095 -0.19154358
#> anansi_ID_sample_9_1  -1.25119684  0.27227702
#> anansi_ID_sample_10_1  0.40026604 -1.08165901
#> 
#> [[31]]
#>                              y_1         x_6
#> anansi_ID_sample_1_1  -1.5727175 -2.32939936
#> anansi_ID_sample_2_1  -0.9864274 -0.54962596
#> anansi_ID_sample_3_1   1.9835123 -0.07257999
#> anansi_ID_sample_4_1  -1.8513784  1.03228840
#> anansi_ID_sample_5_1  -0.9097882  0.21513853
#> anansi_ID_sample_6_1  -1.9514400 -0.49434012
#> anansi_ID_sample_7_1  -0.8002765  1.51213828
#> anansi_ID_sample_8_1  -1.8693616 -0.57946326
#> anansi_ID_sample_9_1  -0.7505715  1.67478963
#> anansi_ID_sample_10_1 -0.5909384 -1.00098026
#> 
#> [[32]]
#>                              y_4         x_6
#> anansi_ID_sample_1_1  -1.0945682 -2.32939936
#> anansi_ID_sample_2_1   0.3648741 -0.54962596
#> anansi_ID_sample_3_1  -1.0080171 -0.07257999
#> anansi_ID_sample_4_1   0.5549964  1.03228840
#> anansi_ID_sample_5_1   0.4360659  0.21513853
#> anansi_ID_sample_6_1   0.1053733 -0.49434012
#> anansi_ID_sample_7_1  -0.2537379  1.51213828
#> anansi_ID_sample_8_1  -0.9764912 -0.57946326
#> anansi_ID_sample_9_1   0.3942178  1.67478963
#> anansi_ID_sample_10_1  1.0149030 -1.00098026
#> 
#> [[33]]
#>                              y_5         x_6
#> anansi_ID_sample_1_1   0.6314438 -2.32939936
#> anansi_ID_sample_2_1  -0.4833983 -0.54962596
#> anansi_ID_sample_3_1  -1.5263339 -0.07257999
#> anansi_ID_sample_4_1   0.6168423  1.03228840
#> anansi_ID_sample_5_1   1.2292868  0.21513853
#> anansi_ID_sample_6_1   0.1574530 -0.49434012
#> anansi_ID_sample_7_1   1.4063421  1.51213828
#> anansi_ID_sample_8_1   0.3968009 -0.57946326
#> anansi_ID_sample_9_1   0.3264230  1.67478963
#> anansi_ID_sample_10_1 -1.4593433 -1.00098026
#> 
#> [[34]]
#>                               y_8         x_6
#> anansi_ID_sample_1_1   1.01340453 -2.32939936
#> anansi_ID_sample_2_1   1.05360525 -0.54962596
#> anansi_ID_sample_3_1  -0.07142864 -0.07257999
#> anansi_ID_sample_4_1   0.95244461  1.03228840
#> anansi_ID_sample_5_1  -0.38403752  0.21513853
#> anansi_ID_sample_6_1  -0.41775999 -0.49434012
#> anansi_ID_sample_7_1  -0.47026817  1.51213828
#> anansi_ID_sample_8_1  -1.69018029 -0.57946326
#> anansi_ID_sample_9_1  -0.94549766  1.67478963
#> anansi_ID_sample_10_1 -0.34301196 -1.00098026
#> 
#> [[35]]
#>                             y_10         x_6
#> anansi_ID_sample_1_1   0.2450964 -2.32939936
#> anansi_ID_sample_2_1  -0.7793198 -0.54962596
#> anansi_ID_sample_3_1   0.5949239 -0.07257999
#> anansi_ID_sample_4_1   1.1088909  1.03228840
#> anansi_ID_sample_5_1  -0.9429690  0.21513853
#> anansi_ID_sample_6_1   0.7003440 -0.49434012
#> anansi_ID_sample_7_1  -0.4246748  1.51213828
#> anansi_ID_sample_8_1  -1.1431376 -0.57946326
#> anansi_ID_sample_9_1   0.2334402  1.67478963
#> anansi_ID_sample_10_1 -0.1892366 -1.00098026
#> 
#> [[36]]
#>                             y_11         x_6
#> anansi_ID_sample_1_1  -1.6634113 -2.32939936
#> anansi_ID_sample_2_1   1.9156908 -0.54962596
#> anansi_ID_sample_3_1  -0.8168944 -0.07257999
#> anansi_ID_sample_4_1   0.3836505  1.03228840
#> anansi_ID_sample_5_1  -0.4586323  0.21513853
#> anansi_ID_sample_6_1  -0.7157040 -0.49434012
#> anansi_ID_sample_7_1   0.4295485  1.51213828
#> anansi_ID_sample_8_1  -1.1698527 -0.57946326
#> anansi_ID_sample_9_1   1.3220883  1.67478963
#> anansi_ID_sample_10_1  1.3659646 -1.00098026
#> 
#> [[37]]
#>                               y_2        x_7
#> anansi_ID_sample_1_1  -0.74209927  1.2227028
#> anansi_ID_sample_2_1   0.69351208  1.0771882
#> anansi_ID_sample_3_1  -0.05946397 -0.6119455
#> anansi_ID_sample_4_1  -1.86389510  0.5068670
#> anansi_ID_sample_5_1  -1.27450892  0.4600684
#> anansi_ID_sample_6_1  -1.78177081  1.4843919
#> anansi_ID_sample_7_1  -0.50859339  0.8819632
#> anansi_ID_sample_8_1  -1.73687829 -0.5367022
#> anansi_ID_sample_9_1   0.04048130  1.2855537
#> anansi_ID_sample_10_1 -0.12417946  0.5878485
#> 
#> [[38]]
#>                              y_3        x_7
#> anansi_ID_sample_1_1  -0.6127534  1.2227028
#> anansi_ID_sample_2_1   0.1606625  1.0771882
#> anansi_ID_sample_3_1  -0.6623596 -0.6119455
#> anansi_ID_sample_4_1  -0.3348517  0.5068670
#> anansi_ID_sample_5_1   0.6230116  0.4600684
#> anansi_ID_sample_6_1   1.0281943  1.4843919
#> anansi_ID_sample_7_1  -1.1345782  0.8819632
#> anansi_ID_sample_8_1   0.9169111 -0.5367022
#> anansi_ID_sample_9_1   1.2138748  1.2855537
#> anansi_ID_sample_10_1 -0.6859325  0.5878485
#> 
#> [[39]]
#>                              y_4        x_7
#> anansi_ID_sample_1_1  -1.0945682  1.2227028
#> anansi_ID_sample_2_1   0.3648741  1.0771882
#> anansi_ID_sample_3_1  -1.0080171 -0.6119455
#> anansi_ID_sample_4_1   0.5549964  0.5068670
#> anansi_ID_sample_5_1   0.4360659  0.4600684
#> anansi_ID_sample_6_1   0.1053733  1.4843919
#> anansi_ID_sample_7_1  -0.2537379  0.8819632
#> anansi_ID_sample_8_1  -0.9764912 -0.5367022
#> anansi_ID_sample_9_1   0.3942178  1.2855537
#> anansi_ID_sample_10_1  1.0149030  0.5878485
#> 
#> [[40]]
#>                               y_6        x_7
#> anansi_ID_sample_1_1  -0.79929553  1.2227028
#> anansi_ID_sample_2_1   0.47001280  1.0771882
#> anansi_ID_sample_3_1  -1.15018052 -0.6119455
#> anansi_ID_sample_4_1  -0.27159672  0.5068670
#> anansi_ID_sample_5_1   0.45742422  0.4600684
#> anansi_ID_sample_6_1  -0.01695794  1.4843919
#> anansi_ID_sample_7_1  -0.54159436  0.8819632
#> anansi_ID_sample_8_1   0.87581342 -0.5367022
#> anansi_ID_sample_9_1   0.74083815  1.2855537
#> anansi_ID_sample_10_1 -0.16481239  0.5878485
#> 
#> [[41]]
#>                             y_11        x_7
#> anansi_ID_sample_1_1  -1.6634113  1.2227028
#> anansi_ID_sample_2_1   1.9156908  1.0771882
#> anansi_ID_sample_3_1  -0.8168944 -0.6119455
#> anansi_ID_sample_4_1   0.3836505  0.5068670
#> anansi_ID_sample_5_1  -0.4586323  0.4600684
#> anansi_ID_sample_6_1  -0.7157040  1.4843919
#> anansi_ID_sample_7_1   0.4295485  0.8819632
#> anansi_ID_sample_8_1  -1.1698527 -0.5367022
#> anansi_ID_sample_9_1   1.3220883  1.2855537
#> anansi_ID_sample_10_1  1.3659646  0.5878485
#> 
#> [[42]]
#>                              y_12        x_7
#> anansi_ID_sample_1_1  -0.24972371  1.2227028
#> anansi_ID_sample_2_1  -0.42472135  1.0771882
#> anansi_ID_sample_3_1   1.09843934 -0.6119455
#> anansi_ID_sample_4_1   0.67092304  0.5068670
#> anansi_ID_sample_5_1   0.03095466  0.4600684
#> anansi_ID_sample_6_1   2.36851196  1.4843919
#> anansi_ID_sample_7_1   2.58578811  0.8819632
#> anansi_ID_sample_8_1   1.21228095 -0.5367022
#> anansi_ID_sample_9_1  -1.25119684  1.2855537
#> anansi_ID_sample_10_1  0.40026604  0.5878485
#> 
#> [[43]]
#>                              y_1        x_8
#> anansi_ID_sample_1_1  -1.5727175 -1.3084820
#> anansi_ID_sample_2_1  -0.9864274  0.3167263
#> anansi_ID_sample_3_1   1.9835123  1.1941583
#> anansi_ID_sample_4_1  -1.8513784  0.9130571
#> anansi_ID_sample_5_1  -0.9097882 -0.7867530
#> anansi_ID_sample_6_1  -1.9514400 -0.4109297
#> anansi_ID_sample_7_1  -0.8002765  0.4763737
#> anansi_ID_sample_8_1  -1.8693616  0.1859159
#> anansi_ID_sample_9_1  -0.7505715 -1.3247163
#> anansi_ID_sample_10_1 -0.5909384  1.1723711
#> 
#> [[44]]
#>                              y_4        x_8
#> anansi_ID_sample_1_1  -1.0945682 -1.3084820
#> anansi_ID_sample_2_1   0.3648741  0.3167263
#> anansi_ID_sample_3_1  -1.0080171  1.1941583
#> anansi_ID_sample_4_1   0.5549964  0.9130571
#> anansi_ID_sample_5_1   0.4360659 -0.7867530
#> anansi_ID_sample_6_1   0.1053733 -0.4109297
#> anansi_ID_sample_7_1  -0.2537379  0.4763737
#> anansi_ID_sample_8_1  -0.9764912  0.1859159
#> anansi_ID_sample_9_1   0.3942178 -1.3247163
#> anansi_ID_sample_10_1  1.0149030  1.1723711
#> 
#> [[45]]
#>                              y_5        x_8
#> anansi_ID_sample_1_1   0.6314438 -1.3084820
#> anansi_ID_sample_2_1  -0.4833983  0.3167263
#> anansi_ID_sample_3_1  -1.5263339  1.1941583
#> anansi_ID_sample_4_1   0.6168423  0.9130571
#> anansi_ID_sample_5_1   1.2292868 -0.7867530
#> anansi_ID_sample_6_1   0.1574530 -0.4109297
#> anansi_ID_sample_7_1   1.4063421  0.4763737
#> anansi_ID_sample_8_1   0.3968009  0.1859159
#> anansi_ID_sample_9_1   0.3264230 -1.3247163
#> anansi_ID_sample_10_1 -1.4593433  1.1723711
#> 
#> [[46]]
#>                               y_8        x_8
#> anansi_ID_sample_1_1   1.01340453 -1.3084820
#> anansi_ID_sample_2_1   1.05360525  0.3167263
#> anansi_ID_sample_3_1  -0.07142864  1.1941583
#> anansi_ID_sample_4_1   0.95244461  0.9130571
#> anansi_ID_sample_5_1  -0.38403752 -0.7867530
#> anansi_ID_sample_6_1  -0.41775999 -0.4109297
#> anansi_ID_sample_7_1  -0.47026817  0.4763737
#> anansi_ID_sample_8_1  -1.69018029  0.1859159
#> anansi_ID_sample_9_1  -0.94549766 -1.3247163
#> anansi_ID_sample_10_1 -0.34301196  1.1723711
#> 
#> [[47]]
#>                               y_9        x_8
#> anansi_ID_sample_1_1   0.35846241 -1.3084820
#> anansi_ID_sample_2_1   0.04818886  0.3167263
#> anansi_ID_sample_3_1   1.13017228  1.1941583
#> anansi_ID_sample_4_1  -0.52763231  0.9130571
#> anansi_ID_sample_5_1   0.47957533 -0.7867530
#> anansi_ID_sample_6_1   1.41752930 -0.4109297
#> anansi_ID_sample_7_1  -0.08009077  0.4763737
#> anansi_ID_sample_8_1  -0.39418700  0.1859159
#> anansi_ID_sample_9_1   0.52525764 -1.3247163
#> anansi_ID_sample_10_1 -0.79593510  1.1723711
#> 
#> [[48]]
#>                             y_11        x_8
#> anansi_ID_sample_1_1  -1.6634113 -1.3084820
#> anansi_ID_sample_2_1   1.9156908  0.3167263
#> anansi_ID_sample_3_1  -0.8168944  1.1941583
#> anansi_ID_sample_4_1   0.3836505  0.9130571
#> anansi_ID_sample_5_1  -0.4586323 -0.7867530
#> anansi_ID_sample_6_1  -0.7157040 -0.4109297
#> anansi_ID_sample_7_1   0.4295485  0.4763737
#> anansi_ID_sample_8_1  -1.1698527  0.1859159
#> anansi_ID_sample_9_1   1.3220883 -1.3247163
#> anansi_ID_sample_10_1  1.3659646  1.1723711
#> 

pairwiseApply(
    X = web,
    FUN = function(x, y) cor(x, y),
    MoreArgs = NULL, SIMPLIFY = TRUE, USE.NAMES = TRUE
)
#>      x_1y_1      x_1y_3      x_1y_9     x_1y_11     x_1y_12      x_2y_1 
#>  0.52003483 -0.83581439 -0.29089933 -0.25181791  0.20467961  0.33936696 
#>      x_2y_2      x_2y_5      x_2y_7      x_2y_8      x_2y_9     x_2y_11 
#>  0.50706667 -0.03719984 -0.48098689  0.27804308  0.30358195 -0.07804186 
#>      x_3y_2      x_3y_3      x_3y_4      x_3y_7      x_3y_8     x_3y_10 
#> -0.17989205 -0.06219585  0.01848036  0.23520804 -0.13339794  0.10857588 
#>     x_3y_12      x_4y_1      x_4y_2      x_4y_3      x_4y_6      x_4y_8 
#>  0.43118276 -0.43591619  0.25811678  0.29317273  0.50650818  0.22154368 
#>      x_4y_9     x_4y_12      x_5y_6      x_5y_9     x_5y_10     x_5y_12 
#> -0.25940303 -0.26446353 -0.43668539  0.16272488  0.06106549  0.11443504 
#>      x_6y_1      x_6y_4      x_6y_5      x_6y_8     x_6y_10     x_6y_11 
#>  0.12474860  0.34407564  0.33523294 -0.29010171  0.11676505  0.43192110 
#>      x_7y_2      x_7y_3      x_7y_4      x_7y_6     x_7y_11     x_7y_12 
#>  0.14425741  0.14704068  0.41151217  0.10078101  0.32832895 -0.17747397 
#>      x_8y_1      x_8y_4      x_8y_5      x_8y_8      x_8y_9     x_8y_11 
#>  0.39615518  0.13188373 -0.54651494  0.08179128 -0.38949748  0.25270138