S4 Methods for MultiFactor

a set of methods to work with MultiFactor, a special type of list.

getEdgeList: Return a data frame in edge list format.

dim: Display a vector of dims (n objects, n ids).

names: Display a vector of dims (n objects, n ids).

dimnames: Display a vector of dims (n objects, n ids).

Extract all elements that contain a given feature type.

[[<- should be used to replace, not add data.

Combine two or more MultiFactor objects.

show: Display the object

Analogous to factors. unfactor(MultiFactor) returns a named list with character data frames with same dimensions as input.

Analogous to factors. droplevels(MultiFactor) returns a MultiFactor with unused levels removed.

get object levels

get object map

Usage

# S4 method for class 'MultiFactor'
getEdgeList(x)

# S4 method for class 'MultiFactor'
dim(x)

# S4 method for class 'MultiFactor'
names(x)

# S4 method for class 'MultiFactor'
dimnames(x)

# S4 method for class 'MultiFactor,ANY,ANY'
x[[i, j, ...]]

# S4 method for class 'MultiFactor,ANY,ANY'
x[[i, j, ...]] <- value

# S4 method for class 'MultiFactor'
c(x, ...)

# S4 method for class 'MultiFactor'
show(object)

# S4 method for class 'MultiFactor'
unfactor(x)

droplevels.MultiFactor(x, exclude = NULL, select = NULL, ...)

# S4 method for class 'MultiFactor'
droplevels(x, ...)

levels.MultiFactor(x)

# S4 method for class 'MultiFactor'
levels(x)

# S4 method for class 'MultiFactor'
levels(x) <- value

# S4 method for class 'MultiFactor'
dictionary(x, ...)

# S4 method for class 'MultiFactor'
dictionary(x, ...) <- value

# S4 method for class 'MultiFactor'
subset(x, subset, select, ...)

Arguments

x

MultiFactor object

...

Optional metadata columns.

value

a replacement character vector of suitable dimensions.

object

Any R object

exclude

NULL or Named character list of similar structure as levels(MultiFactor). Which levels to drop from output.

select

expression. Which column names to consider. If missing (Default), consider all column names.

subset

logical expression indicating rows to keep. Must contain variables found as column names.

Value

A named character list

A MultiFactor

a named list of character vectors.

a named sparse biadjacency matrix of dimensions (dimnames(x))

Details

if argument i is a vector, return all elements that contain at least that exact combination.

Only one of select and exclude should be provided, as they are each others complement.

subset: For MultiFactor objects, sub-setting is only applied to data frames compatible with the expression. The rest are returned unaltered. Modeled after subset().

See also

Examples

# prep input
l <- asMultiFactor(kegg_link())

# Sub-setting is only performed on data frames that contain the arguments
str(subset(x = l, cpd %in% c("C00001", "C00002")))
#> Formal class 'MultiFactor' [package "anansi"] with 3 slots
#>   ..@ index :List of 2
#>   .. ..$ ec2ko :'data.frame':	8156 obs. of  2 variables:
#>   .. .. ..$ ec: int [1:8156] 1 2 3 4 5 6 7 8 9 10 ...
#>   .. .. ..$ ko: int [1:8156] 1 2 3 4 4 5 6 7 8 9 ...
#>   .. ..$ ec2cpd:'data.frame':	0 obs. of  2 variables:
#>   .. .. ..$ ec : int(0) 
#>   .. .. ..$ cpd: int(0) 
#>   ..@ levels:List of 3
#>   .. ..$ ec : chr [1:4825] "1.1.1.1" "1.1.1.2" "1.1.1.3" "1.1.1.4" ...
#>   .. ..$ ko : chr [1:7270] "K00001" "K00002" "K00003" "K00004" ...
#>   .. ..$ cpd: chr [1:6754] "C00001" "C00002" "C00003" "C00004" ...
#>   ..@ map   :Formal class 'dgCMatrix' [package "Matrix"] with 6 slots
#>   .. .. ..@ i       : int [1:4] 0 1 0 1
#>   .. .. ..@ p       : int [1:4] 0 2 3 4
#>   .. .. ..@ Dim     : int [1:2] 2 3
#>   .. .. ..@ Dimnames:List of 2
#>   .. .. .. ..$ : chr [1:2] "ec2ko" "ec2cpd"
#>   .. .. .. ..$ : chr [1:3] "ec" "ko" "cpd"
#>   .. .. ..@ x       : num [1:4] 4825 4825 7270 6754
#>   .. .. ..@ factors : list()

# Several data frames at the same time:
subset(x = l, ec %in% c("1.2.3.4", "4.3.2.1"))
#> A list of class MultiFactor,
#>     3 feature types across 2 edge lists.
#> 
#>          ec   ko  cpd
#> ec2ko  4825 7270    .
#> ec2cpd 4825    . 6754
#> 
#> Values represent unique feature names in that edge list.
#> 
#> Levels:
#> 
#> ec  : 4825 Levels: 1.1.1.1 1.1.1.2 ... 1.14.13.253 
#> ko  : 7270 Levels: K00001 K00002 ... K28089 
#> cpd : 6754 Levels: C00001 C00002 ... C22998