Calculate an association network

This is the main workspider function in the anansi package. It manages the individual functionalities of anansi, including correlation analysis, correlation by group and differential correlation.

Usage

anansi(
  web,
  formula = ~1,
  groups = NULL,
  metadata,
  adjust.method = "BH",
  verbose = TRUE,
  ignore_dictionary = FALSE,
  return.format = "table"
)

Arguments

web

An anansiWeb object, containing two tables with 'omics data and a dictionary that links them. See weaveWebFromTables() for how to weave a web.

formula

A formula object. Used to assess differential associations.

groups

A vector of the column names of categorical values in the metadata object to control which groups should be assessed for simple correlations. If no argument provided, anansi will let you know and still to regular correlations according to your dictionary.

metadata

A vector or data.frame of categorical or continuous value necessary for differential correlations. Typically a state or treatment. If no argument provided, anansi will let you know and still to regular correlations according to your dictionary.

adjust.method

Method to adjust p-values for multiple comparisons. adjust.method = "BH" is the default value. See p.adjust() in the base R stats package.

verbose

A boolean. Toggles whether to print diagnostic information while running. Useful for debugging errors on large datasets.

ignore_dictionary

A boolean. Default is FALSE. If set to TRUE, regular all vs all associations will be tested regardless of the dictionary.

return.format

Character scalar. Should be one of "table" , "list", or "raw". Should the output of anansi respectively be a wide `data.frame` of results, a list containing the results and input, or a list of raw output (used for testing purposes). convenient use. (Default: "table")

Value

A list of lists containing correlation coefficients, p-values and q-values for all operations.

Examples

# Load example data:

data(dictionary)
data(FMT_data)

# Clean and prepare the example data.
# In the example dataset, the metabolites are already cleaned.

KOs <- floor(FMT_KOs)
KOs <- apply(KOs, c(1, 2), function(x) as.numeric(as.character(x)))
KOs <- KOs[apply(KOs == 0, 1, sum) <= (ncol(KOs) * 0.90), ]

KOs <- KOs[row.names(KOs) %in% sort(unique(unlist(anansi_dic))), ]

# CLR-transform.

KOs.exp <- clr_c(KOs)

# Make sure that columns are features and rows are samples.

t1 <- t(FMT_metab)
t2 <- t(KOs.exp)

# Run anansi pipeline.

web <- weaveWebFromTables(
  tableY = t1,
  tableX = t2,
  dictionary = anansi_dic
)
#> Operating in interaction mode
#> 3 were matched between table 1 and the columns of the adjacency matrix
#> 50 were matched between table 2 and the rows of the adjacency matrix

anansi_out <- anansi(
  web = web,
  formula = ~Legend,
  groups = "Legend",
  metadata = FMT_metadata,
  adjust.method = "BH",
  verbose = TRUE
)
#> Fitting least-squares for following model:
#> ~ x + Legend + x:Legend 
#> Running correlations for the following groups:
#>  Aged yFMT, Aged oFMT, Young yFMT

# Only consider interactions where the entire model fits well enough.
anansiLong <- anansiLong[anansiLong$full_q.values < 0.2, ]
#> Error: object 'anansiLong' not found

library(ggplot2)

p <- plotAnansi(anansi_out,
                association.type = "disjointed",
                model.var = "Legend",
                fill_by = "group",
                signif.threshold = 0.05,
                x_lab = "Pearson's rho")
#> Error in mutate(separate_wider_regex(pivot_wider(separate_wider_regex(pivot_longer(x,     matches(paste0("^", patterns[["All"]]))), name, c(group = patterns[["All"]],     cor_param = ".\\.values$")), names_from = cor_param, values_from = value),     group, unlist(patterns[-1]), too_few = "align_start", cols_remove = FALSE),     across(all_of(group_cols), ~gsub("_$", "", replace_na(.,         "All")))): could not find function "mutate"
p + scale_fill_manual(values = c(
    "Young yFMT" = "#2166ac",
    "Aged oFMT" = "#b2182b",
    "Aged yFMT" = "#ef8a62",
    "All" = "gray"
  )) +
  theme_bw()
#> Error: object 'p' not found

# See also ?spinToPlots