S4 Methods for anansiLinkMap

a set of methods to work with anansiLinkWeb, a special type of list.

names: Display a list of column names from anansiLinkMap

show: Display the object

getEdgeList: Return a data frame in edge list format.

Usage

# S4 method for class 'anansiLinkMap'
names(x)

# S4 method for class 'anansiLinkMap'
show(object)

# S4 method for class 'anansiLinkMap'
getEdgeList(x)

# S4 method for class 'anansiLinkMap'
subset(x, subset, select, ...)

Arguments

x

A vector-like, matrix-like or data-frame-like object to be subsetted.

object

Any R object

subset

logical expression indicating rows to keep. Must contain variables found as column names.

select

expression. Which column names to consider. If missing (Default), consider all column names.

...

Additional arguments (e.g. subset, select, drop), for use in specific methods. See ?base::subset for more information.

Details

subset: For anansiLinkMap objects, sub-setting is only applied to data frames compatible with the expression. The rest are returned unaltered. Modeled after subset().

See also

Examples

# prep input
l <- asLinkMap(kegg_link())

# Sub-setting is only performed on data frames that contain the arguments
str(subset(x = l, cpd %in% c("C00001", "C00002")))
#> Formal class 'anansiLinkMap' [package "anansi"] with 1 slot
#>   ..@ .Data:List of 2
#>   .. ..$ :'data.frame':	4216 obs. of  2 variables:
#>   .. .. ..$ ec: chr [1:4216] "1.1.1.1" "1.1.1.22" "1.1.1.23" "1.1.1.132" ...
#>   .. .. ..$ ko: chr [1:4216] "K00001" "K00012" "K00013" "K00066" ...
#>   .. ..$ :'data.frame':	2235 obs. of  2 variables:
#>   .. .. ..$ ec : chr [1:2235] "1.1.1.1" "1.1.1.132" "1.1.1.136" "1.1.1.170" ...
#>   .. .. ..$ cpd: chr [1:2235] "C00001" "C00001" "C00001" "C00001" ...
#>   ..$ names: chr [1:2] "ec2ko" "ec2cpd"

# Several data frames at the same time:
subset(x = l, ec %in% c("1.2.3.4", "4.3.2.1"))
#> An object of class anansiLinkMap:
#> $ec2ko
#> ec ko 
#>  1  2 
#> 
#> $ec2cpd
#>  ec cpd 
#>   1   3 
#>