Produce a biadjacency matrix given tables and a dictionary

calculates a biadjacency matrix for the cases where link is a single - or a list of two - data.frame(s).

Usage

deliver_web_cases(link, terms, tableX, tableY, link_is_list)

Arguments

link

a data.frame or list of two compatible ones.

terms

a length 2 character vector, naming the x and y terms in order

tableX

A table containing features of interest. Rows should be samples and columns should be features. The Y and X refer to the position of the features in a formula: Y ~ X.

tableY

A table containing features of interest. Rows should be samples and columns should be features. The Y and X refer to the position of the features in a formula: Y ~ X.

link_is_list

a boolean, is a list (or FALSE: a data.frame)

Value

a sparse boolean biadjacency matrix, for use in main anansi workflow