anansi wrapper for the MultiAssayExperiment class

getAnansi provides a wrapper to execute the anansi pipeline on a MultiAssayExperiment::MultiAssayExperiment() object. It applies the functions weaveWeb() and anansi() in the given order.

Usage

getAnansi(x, ...)

# S4 method for class 'MultiAssayExperiment'
getAnansi(
  x,
  formula,
  link = NULL,
  tableY = NULL,
  tableX = NULL,
  ...,
  force_new = FALSE
)

Arguments

x

a MultiAssayExperiment::MultiAssayExperiment().

...

additional parameters that can be passed to AnansiWeb() or anansi().

formula

A formula object. Used to assess differential associations.

link

One of the following:

Character scalar with value "none"

data.frame with two columns

list with two such data.frames

tableY, tableX

Character scalar or numeric scalar. Selects experiment corresponding to tableY and tableX from experiments(x) of MultiAssayExperiment object by name or index, name is recommended. (Default slots: Y = 1L, X = 2L).

force_new

boolean If x already has a dictionary Matrix in metadata, ignore it and generate a new object anyway? (Default: FALSE).

Value

If return.format is "table" (default), a wide format data.frame intended to be compatible with ggplot2, or specialized plotting functions (See plotAnansi()). If return.format is "list", a list with aforementioned table, as well as input and additional information. If return.format is "raw", a list of raw output (used for testing purposes).

Details

This wrapper of anansi() allows to perform a complete anansi analysis directly on objects of class MultiAssayExperiment::MultiAssayExperiment(). First, the assays from experiments specified by tableX and tableY are passed to AnansiWeb() to build an AnansiWeb object. If there are more than one assay in that experiment, the specific assay can by specified using assay.type1 and assay.type2. Next, this object is fed to the main anansi() function to compute interactions between the two assays.

See also

Examples

# Import libraries
library(mia)
#> Loading required package: MultiAssayExperiment
#> Loading required package: SummarizedExperiment
#> Loading required package: MatrixGenerics
#> Loading required package: matrixStats
#> 
#> Attaching package: ‘MatrixGenerics’
#> The following objects are masked from ‘package:matrixStats’:
#> 
#>     colAlls, colAnyNAs, colAnys, colAvgsPerRowSet, colCollapse,
#>     colCounts, colCummaxs, colCummins, colCumprods, colCumsums,
#>     colDiffs, colIQRDiffs, colIQRs, colLogSumExps, colMadDiffs,
#>     colMads, colMaxs, colMeans2, colMedians, colMins, colOrderStats,
#>     colProds, colQuantiles, colRanges, colRanks, colSdDiffs, colSds,
#>     colSums2, colTabulates, colVarDiffs, colVars, colWeightedMads,
#>     colWeightedMeans, colWeightedMedians, colWeightedSds,
#>     colWeightedVars, rowAlls, rowAnyNAs, rowAnys, rowAvgsPerColSet,
#>     rowCollapse, rowCounts, rowCummaxs, rowCummins, rowCumprods,
#>     rowCumsums, rowDiffs, rowIQRDiffs, rowIQRs, rowLogSumExps,
#>     rowMadDiffs, rowMads, rowMaxs, rowMeans2, rowMedians, rowMins,
#>     rowOrderStats, rowProds, rowQuantiles, rowRanges, rowRanks,
#>     rowSdDiffs, rowSds, rowSums2, rowTabulates, rowVarDiffs, rowVars,
#>     rowWeightedMads, rowWeightedMeans, rowWeightedMedians,
#>     rowWeightedSds, rowWeightedVars
#> Loading required package: GenomicRanges
#> Loading required package: stats4
#> Loading required package: BiocGenerics
#> Loading required package: generics
#> 
#> Attaching package: ‘generics’
#> The following objects are masked from ‘package:base’:
#> 
#>     as.difftime, as.factor, as.ordered, intersect, is.element, setdiff,
#>     setequal, union
#> 
#> Attaching package: ‘BiocGenerics’
#> The following objects are masked from ‘package:stats’:
#> 
#>     IQR, mad, sd, var, xtabs
#> The following objects are masked from ‘package:base’:
#> 
#>     Filter, Find, Map, Position, Reduce, anyDuplicated, aperm, append,
#>     as.data.frame, basename, cbind, colnames, dirname, do.call,
#>     duplicated, eval, evalq, get, grep, grepl, is.unsorted, lapply,
#>     mapply, match, mget, order, paste, pmax, pmax.int, pmin, pmin.int,
#>     rank, rbind, rownames, sapply, saveRDS, table, tapply, unique,
#>     unsplit, which.max, which.min
#> Loading required package: S4Vectors
#> 
#> Attaching package: ‘S4Vectors’
#> The following object is masked from ‘package:tidyr’:
#> 
#>     expand
#> The following object is masked from ‘package:utils’:
#> 
#>     findMatches
#> The following objects are masked from ‘package:base’:
#> 
#>     I, expand.grid, unname
#> Loading required package: IRanges
#> Loading required package: Seqinfo
#> Loading required package: Biobase
#> Welcome to Bioconductor
#> 
#>     Vignettes contain introductory material; view with
#>     'browseVignettes()'. To cite Bioconductor, see
#>     'citation("Biobase")', and for packages 'citation("pkgname")'.
#> 
#> Attaching package: ‘Biobase’
#> The following object is masked from ‘package:MatrixGenerics’:
#> 
#>     rowMedians
#> The following objects are masked from ‘package:matrixStats’:
#> 
#>     anyMissing, rowMedians
#> Loading required package: SingleCellExperiment
#> Loading required package: TreeSummarizedExperiment
#> Loading required package: Biostrings
#> Loading required package: XVector
#> 
#> Attaching package: ‘Biostrings’
#> The following object is masked from ‘package:base’:
#> 
#>     strsplit
#> This is mia version 1.17.0
#> - Online documentation and vignettes: https://microbiome.github.io/mia/
#> - Online book 'Orchestrating Microbiome Analysis (OMA)': https://microbiome.github.io/OMA/docs/devel/
library(TreeSummarizedExperiment)
library(MultiAssayExperiment)

web <- randomWeb(n_samples = 100)
mae <- as(web, "MultiAssayExperiment")

# Perform anansi analysis
out <- getAnansi(mae,
    tableY = "y", tableX = "x",
    formula = ~group_ab
)
#> Fitting least-squares for following model:
#> ~ x + group_ab + x:group_ab 
#> Running correlations for the following groups:
#>  a, b

# View subset of results
head(out, 5)
#>   feature_Y feature_X All_r.values All_t.values All_p.values   a_r.values
#> 1       y_1       x_1   0.01864639    0.1846220    0.8539071  0.081443836
#> 2       y_2       x_1  -0.04145129    0.4106998    0.6821893 -0.095129570
#> 3       y_5       x_1   0.09412145    0.9359096    0.3516203  0.005507361
#> 4       y_6       x_1   0.12175923    1.2143903    0.2275169  0.152280579
#> 5       y_7       x_1  -0.12430769    1.2402027    0.2178616 -0.088931257
#>   a_t.values a_p.values  b_r.values b_t.values b_p.values full_r.squared
#> 1 0.56021187  0.5779956 -0.03836262 0.26873618  0.7892610    0.003598460
#> 2 0.65514663  0.5155673  0.00300032 0.02100233  0.9833291    0.005309043
#> 3 0.03775714  0.9700412  0.16096072 1.14161071  0.2591643    0.015240296
#> 4 1.05630238  0.2962296  0.08185820 0.57493687  0.5679646    0.014385564
#> 5 0.61210729  0.5434168 -0.16594610 1.17795524  0.2445039    0.016460067
#>   full_f.values full_p.values disjointed_group_ab_r.squared
#> 1     0.1733499     0.8411060                  0.0033988128
#> 2     0.2561942     0.7745191                  0.0029855167
#> 3     0.7428556     0.4784680                  0.0067149669
#> 4     0.7005854     0.4988141                  0.0012565949
#> 5     0.8033057     0.4508332                  0.0006796926
#>   disjointed_group_ab_f.values disjointed_group_ab_p.values
#> 1                   0.32739879                    0.5685323
#> 2                   0.28746784                    0.5930880
#> 3                   0.64899481                    0.4224632
#> 4                   0.12078489                    0.7289456
#> 5                   0.06529487                    0.7988615
#>   emergent_group_ab_r.squared emergent_group_ab_f.values
#> 1                  0.02004087                   1.942819
#> 2                  0.02004087                   1.942819
#> 3                  0.02004087                   1.942819
#> 4                  0.02004087                   1.942819
#> 5                  0.02004087                   1.942819
#>   emergent_group_ab_p.values All_q.values a_q.values b_q.values full_q.values
#> 1                  0.1666147    0.9901415  0.8207615  0.8810355     0.9391157
#> 2                  0.1666147    0.9896062  0.8207615  0.9833291     0.9391157
#> 3                  0.1666147    0.9896062  0.9819187  0.8754018     0.8682065
#> 4                  0.1666147    0.9187568  0.8207615  0.8810355     0.8682065
#> 5                  0.1666147    0.9187568  0.8207615  0.8754018     0.8682065
#>   disjointed_group_ab_q.values emergent_group_ab_q.values
#> 1                    0.9549765                  0.7270462
#> 2                    0.9549765                  0.7270462
#> 3                    0.9549765                  0.7270462
#> 4                    0.9549765                  0.7270462
#> 5                    0.9832142                  0.7270462