weaveWeb wrapper for the MultiAssayExperiment class

weaveWeb wrapper for the MultiAssayExperiment class

Usage

# S4 method for class 'MultiAssayExperiment'
getWeb(x, tableY = 1, tableX = 2, link = NULL, force_new = FALSE, ...)

Arguments

x

a MultiAssayExperiment::MultiAssayExperiment().

...

additional parameters passed to AnansiWeb().

tableY, tableX

Character scalar or numeric scalar. Selects experiment corresponding to tableY and tableX from experiments(x) of MultiAssayExperiment object by name or index, name is recommended. (Default slots: Y = 1, X = 2).

link

One of the following:

Character scalar with value "none"

data.frame with two columns

list with two such data.frames

force_new

boolean If x already has a dictionary Matrix in metadata, ignore it and generate a new object anyway? (Default: FALSE).

Value

an AnansiWeb object, with sparse binary biadjacency matrix with features from y as rows and features from x as columns in dictionary slot. If x already contains a dictionary in metadata, use that one, unless force_new = TRUE.

Details

This wrapper of weaveWeb() allows to generate an AnansiWeb S4 object directly from objects of class MultiAssayExperiment::MultiAssayExperiment() . First, the assays specified by assay.typeY and assay.typeX are passed to AnansiWeb() to build an AnansiWeb object.

Examples

# Make a random anansiWeb
web <- randomWeb()

# Combine experiments into MultiAssayExperiment object
mae <- as(web, "MultiAssayExperiment")

# Back to AnansiWeb
outWeb <- getWeb(mae, tableY = "y", tableX = "x")