Generate a random AnansiWeb or MultiFactor

Randomly generate a valid AnansiWeb or MultiFactor object.

Usage

randomWeb(
  n_samples = 10,
  n_reps = 1L,
  n_features_x = 8,
  n_features_y = 12,
  sparseness = 0.5,
  tableY = NULL,
  tableX = NULL,
  dictionary = NULL
)

randomMultiFactor(n_types = 6, n_features = 100, sparseness = 0.5)

krebsDemoWeb(n_samples = 100, n_reps = 4L)

Arguments

n_samples, n_reps

Numeric scalar Number of samples and repeated measures of those samples to be generated. Ignored if tableY and tableX are provided. (defaults: 10 samples without repeats)

n_features_y, n_features_x

Numeric scalar Number of features to be generated. Ignored if tableY and tableX are provided.

sparseness

Numeric scalar, proportion: How rare are connections

tableY, tableX

A table containing features of interest. Rows should be samples and columns should be features. Y and X refer to the position of the features in a formula: Y ~ X.

dictionary

A binary adjacency matrix of class Matrix, or coercible to Matrix.

n_types

Numeric scalar, number of types of features to generate

n_features

Numeric scalar, number of features per type

Value

a randomly generated object of the specified class.

See also

Examples

# Make a random AnansiWeb object
randomWeb()
#> <anansi::AnansiWeb>
#>  @ tableY    : num [1:10, 1:12] 0.9051 0.0585 -0.4541 0.6493 -0.5123 ...
#>  .. - attr(*, "dimnames")=List of 2
#>  ..  ..$ sample_id: chr [1:10] "sample_1_1" "sample_2_1" "sample_3_1" "sample_4_1" ...
#>  ..  ..$ y        : chr [1:12] "y_1" "y_2" "y_3" "y_4" ...
#>  @ tableX    : num [1:10, 1:8] 1.026 0.0555 0.5386 0.1539 1.0237 ...
#>  .. - attr(*, "dimnames")=List of 2
#>  ..  ..$ sample_id: chr [1:10] "sample_1_1" "sample_2_1" "sample_3_1" "sample_4_1" ...
#>  ..  ..$ x        : chr [1:8] "x_1" "x_2" "x_3" "x_4" ...
#>  @ dictionary:Formal class 'ngCMatrix' [package "Matrix"] with 5 slots
#>  ..  ..@ i       : int [1:48] 0 2 3 4 7 8 9 0 1 2 ...
#>  ..  ..@ p       : int [1:9] 0 7 15 20 28 35 38 43 48
#>  ..  ..@ Dim     : int [1:2] 12 8
#>  ..  ..@ Dimnames:List of 2
#>  ..  .. ..$ y: chr [1:12] "y_1" "y_2" "y_3" "y_4" ...
#>  ..  .. ..$ x: chr [1:8] "x_1" "x_2" "x_3" "x_4" ...
#>  ..  ..@ factors : list()
#>  @ metadata  :'data.frame':	10 obs. of  7 variables:
#>  .. $ sample_id: chr  "sample_1" "sample_2" "sample_3" "sample_4" ...
#>  .. $ repeated : chr  "rep_1" "rep_1" "rep_1" "rep_1" ...
#>  .. $ group_ab : chr  "a" "a" "b" "b" ...
#>  .. $ subtype  : chr  "x" "z" "z" "y" ...
#>  .. $ score_a  : num  -0.978 -1.135 -0.588 0.497 0.157 ...
#>  .. $ score_b  : num  0.899 -1.925 1.122 -0.438 0.816 ...
#>  .. $ score_c  : num  0.766 -0.143 0.951 0.427 -0.297 ...
krebsDemoWeb()
#> <anansi::AnansiWeb>
#>  @ tableY    : num [1:400, 1:8] -0.903 -1.776 0.401 -0.681 1.351 ...
#>  .. - attr(*, "dimnames")=List of 2
#>  ..  ..$ sample_id: chr [1:400] "sample_1_1" "sample_1_2" "sample_1_3" "sample_1_4" ...
#>  ..  ..$ Enzyme   : chr [1:8] "aconitase" "isocitrate dehydrogenase" "ketoglutarate dehydrogenase" "succinyl-CoA synthetase" ...
#>  @ tableX    : num [1:400, 1:9] -1.123 -1.403 0.57 -0.715 1.456 ...
#>  .. - attr(*, "dimnames")=List of 2
#>  ..  ..$ sample_id : chr [1:400] "sample_1_1" "sample_1_2" "sample_1_3" "sample_1_4" ...
#>  ..  ..$ Metabolite: chr [1:9] "citrate" "cis-aconitate" "isocitrate" "ketoglutarate" ...
#>  @ dictionary:Formal class 'ngCMatrix' [package "Matrix"] with 5 slots
#>  ..  ..@ i       : int [1:17] 0 7 0 0 1 1 2 2 3 3 ...
#>  ..  ..@ p       : int [1:10] 0 2 3 5 7 9 11 13 15 17
#>  ..  ..@ Dim     : int [1:2] 8 9
#>  ..  ..@ Dimnames:List of 2
#>  ..  .. ..$ Enzyme    : chr [1:8] "aconitase" "isocitrate dehydrogenase" "ketoglutarate dehydrogenase" "succinyl-CoA synthetase" ...
#>  ..  .. ..$ Metabolite: chr [1:9] "citrate" "cis-aconitate" "isocitrate" "ketoglutarate" ...
#>  ..  ..@ factors : list()
#>  @ metadata  :'data.frame':	400 obs. of  7 variables:
#>  .. $ sample_id: chr  "sample_1" "sample_1" "sample_1" "sample_1" ...
#>  .. $ repeated : chr  "rep_1" "rep_2" "rep_3" "rep_4" ...
#>  .. $ group_ab : chr  "a" "a" "a" "a" ...
#>  .. $ subtype  : chr  "z" "z" "z" "z" ...
#>  .. $ score_a  : num  -0.723 -0.856 0.115 0.644 1.39 ...
#>  .. $ score_b  : num  -0.259 -0.0341 -0.6469 -0.4338 -0.0847 ...
#>  .. $ score_c  : num  0.246 2.424 2.54 0.698 -0.236 ...
randomMultiFactor()
#> <anansi::MultiFactor>
#>  @ index :List of 5
#>  .. $ a2b:'data.frame':	5000 obs. of  2 variables:
#>  ..  ..$ a: int [1:5000] 1 2 3 4 8 9 10 12 14 15 ...
#>  ..  ..$ b: int [1:5000] 1 1 1 1 1 1 1 1 1 1 ...
#>  .. $ b2c:'data.frame':	5000 obs. of  2 variables:
#>  ..  ..$ b: int [1:5000] 1 3 4 5 10 11 12 13 15 18 ...
#>  ..  ..$ c: int [1:5000] 1 1 1 1 1 1 1 1 1 1 ...
#>  .. $ c2d:'data.frame':	5000 obs. of  2 variables:
#>  ..  ..$ c: int [1:5000] 1 2 3 4 6 7 11 14 16 17 ...
#>  ..  ..$ d: int [1:5000] 1 1 1 1 1 1 1 1 1 1 ...
#>  .. $ d2e:'data.frame':	5000 obs. of  2 variables:
#>  ..  ..$ d: int [1:5000] 1 2 4 6 11 13 14 15 16 21 ...
#>  ..  ..$ e: int [1:5000] 1 1 1 1 1 1 1 1 1 1 ...
#>  .. $ e2f:'data.frame':	5000 obs. of  2 variables:
#>  ..  ..$ e: int [1:5000] 1 2 5 6 7 10 11 12 14 15 ...
#>  ..  ..$ f: int [1:5000] 1 1 1 1 1 1 1 1 1 1 ...
#>  @ levels:List of 6
#>  .. $ a: chr [1:100] "a_001" "a_002" "a_003" "a_004" ...
#>  .. $ b: chr [1:100] "b_001" "b_002" "b_003" "b_004" ...
#>  .. $ c: chr [1:100] "c_001" "c_002" "c_003" "c_004" ...
#>  .. $ d: chr [1:100] "d_001" "d_002" "d_003" "d_004" ...
#>  .. $ e: chr [1:100] "e_001" "e_002" "e_003" "e_004" ...
#>  .. $ f: chr [1:100] "f_001" "f_002" "f_003" "f_004" ...
#>  @ map   :Formal class 'dgCMatrix' [package "Matrix"] with 6 slots
#>  ..  ..@ i       : int [1:10] 0 0 1 1 2 2 3 3 4 4
#>  ..  ..@ p       : int [1:7] 0 1 3 5 7 9 10
#>  ..  ..@ Dim     : int [1:2] 5 6
#>  ..  ..@ Dimnames:List of 2
#>  ..  .. ..$ : chr [1:5] "a2b" "b2c" "c2d" "d2e" ...
#>  ..  .. ..$ : chr [1:6] "a" "b" "c" "d" ...
#>  ..  ..@ x       : num [1:10] 100 100 100 100 100 100 100 100 100 100
#>  ..  ..@ factors : list()