Subset a valid link object

Subset a valid link object

Usage

subset_link(link, subset)

Arguments

link

data.frame with exactly two, named, columns, or a list of several such data.frames

subset

logical expression indicating rows to keep. Must contain variables found as column names.

Details

Convenience function for valid link objects. For link lists, sub-setting is only applied to data frames compatible with the expression. The rest are returned unaltered. Modeled after subset().

See also

Examples

# Works for single link data frame, as well as list of multiple such objects.
head(subset_link(link = ec2cpd, cpd %in% c("C00001", "C00002")))
#>          ec    cpd
#> 1   1.1.1.1 C00001
#> 2 1.1.1.115 C00001
#> 3 1.1.1.132 C00001
#> 4 1.1.1.136 C00001
#> 5 1.1.1.170 C00001
#> 6 1.1.1.186 C00001

# Sub-setting is only performed on data frames that contain the arguments
str(subset_link(link = kegg_link(), cpd %in% c("C00001", "C00002")))
#> List of 2
#>  $ ec2ko :'data.frame':	9792 obs. of  2 variables:
#>   ..$ ec: chr [1:9792] "1.1.1.1" "1.1.1.2" "1.1.1.3" "1.1.1.4" ...
#>   ..$ ko: chr [1:9792] "K00001" "K00002" "K00003" "K00004" ...
#>  $ ec2cpd:'data.frame':	2676 obs. of  2 variables:
#>   ..$ ec : chr [1:2676] "1.1.1.1" "1.1.1.115" "1.1.1.132" "1.1.1.136" ...
#>   ..$ cpd: chr [1:2676] "C00001" "C00001" "C00001" "C00001" ...

# Several data frames at the same time:
str(subset_link(link = kegg_link(), ec %in% c("1.2.3.4", "4.3.2.1")))
#> List of 2
#>  $ ec2ko :'data.frame':	2 obs. of  2 variables:
#>   ..$ ec: chr [1:2] "4.3.2.1" "4.3.2.1"
#>   ..$ ko: chr [1:2] "K01755" "K14681"
#>  $ ec2cpd:'data.frame':	7 obs. of  2 variables:
#>   ..$ ec : chr [1:7] "1.2.3.4" "1.2.3.4" "1.2.3.4" "4.3.2.1" ...
#>   ..$ cpd: chr [1:7] "C00007" "C00011" "C00027" "C00062" ...