Create an anansiWeb object from two 'omics tables and a dictionary

This function will take two tables of 'omics data, for example metabolomics and functional microbiome data. It will also take a dictionary list, which is provided in this package.

Usage

weaveWebFromTables(
  tableY,
  tableX = NULL,
  dictionary = anansi::anansi_dic,
  verbose = TRUE,
  mode = "interaction",
  prune = FALSE,
  max_sds = 3
)

Arguments

tableY

A table containing features of interest. Rows should be samples and columns should be features. The Y and X refer to the position of the features in a formula: Y ~ X.

tableX

A table containing features of interest. Rows should be samples and columns should be features. The Y and X refer to the position of the features in a formula: Y ~ X. If left empty, tableY will be duplicated.

dictionary

A list that has feature names from tableY as names. Default is the dictionary provided in this package.

verbose

A boolean. Toggles whether to print diagnostic information while running. Useful for debugging errors on large datasets.

mode

A character vector. Can be "interaction" or "membership". Toggles whether to link two datasets based on their interactions or based on shared group membership.

prune

A boolean. Toggles whether to prune particularly large groups.

max_sds

A numeric. Only relevant for prune == TRUE. How many SDs larger than the median a group can be before it is pruned. For general use, we recommend sticking to that one. You can access the dictionary like this: data(dictionary)

Value

an anansiWeb object. Web is used as input for most of the main workflow of anansi.