anansi wrapper for the MultiAssayExperiment class

getAnansi provides a wrapper to execute the anansi pipeline on a MultiAssayExperiment::MultiAssayExperiment() object. It applies the functions weaveWeb() and anansi() in the given order.

Usage

getAnansi(x, ...)

# S4 method for class 'MultiAssayExperiment'
getAnansi(
  x,
  tableY = 1,
  tableX = 2,
  formula,
  link = NULL,
  force_new = FALSE,
  ...
)

Arguments

x

a MultiAssayExperiment::MultiAssayExperiment().

...

additional parameters that can be passed to AnansiWeb() or anansi().

tableY, tableX

Character scalar or numeric scalar. Selects experiment corresponding to tableY and tableX from experiments(x) of MultiAssayExperiment object by name or index, name is recommended. (Default slots: Y = 1, X = 2).

formula

A formula object. Used to assess differential associations.

link

One of the following:

Character scalar with value "none"

data.frame with two columns

list with two such data.frames

force_new

boolean If x already has a dictionary Matrix in metadata, ignore it and generate a new object anyway? (Default: FALSE).

Value

If return.format is "table" (default), a wide format data.frame intended to be compatible with ggplot2, or specialized plotting functions (See plotAnansi()). If return.format is "list", a list with aforementioned table, as well as input and additional information. If return.format is "raw", a list of raw output (used for testing purposes).

Details

This wrapper of anansi() allows to perform a complete anansi analysis directly on objects of class MultiAssayExperiment::MultiAssayExperiment() . First, the assays specified by assay.typeY and assay.typeX are passed to AnansiWeb() to build an AnansiWeb object. Next, this object is fed to the main anansi() function to compute interactions between the two assays.

See also

Examples

# Import libraries
library(mia)
#> Loading required package: MultiAssayExperiment
#> Loading required package: SummarizedExperiment
#> Loading required package: MatrixGenerics
#> Loading required package: matrixStats
#> 
#> Attaching package: ‘MatrixGenerics’
#> The following objects are masked from ‘package:matrixStats’:
#> 
#>     colAlls, colAnyNAs, colAnys, colAvgsPerRowSet, colCollapse,
#>     colCounts, colCummaxs, colCummins, colCumprods, colCumsums,
#>     colDiffs, colIQRDiffs, colIQRs, colLogSumExps, colMadDiffs,
#>     colMads, colMaxs, colMeans2, colMedians, colMins, colOrderStats,
#>     colProds, colQuantiles, colRanges, colRanks, colSdDiffs, colSds,
#>     colSums2, colTabulates, colVarDiffs, colVars, colWeightedMads,
#>     colWeightedMeans, colWeightedMedians, colWeightedSds,
#>     colWeightedVars, rowAlls, rowAnyNAs, rowAnys, rowAvgsPerColSet,
#>     rowCollapse, rowCounts, rowCummaxs, rowCummins, rowCumprods,
#>     rowCumsums, rowDiffs, rowIQRDiffs, rowIQRs, rowLogSumExps,
#>     rowMadDiffs, rowMads, rowMaxs, rowMeans2, rowMedians, rowMins,
#>     rowOrderStats, rowProds, rowQuantiles, rowRanges, rowRanks,
#>     rowSdDiffs, rowSds, rowSums2, rowTabulates, rowVarDiffs, rowVars,
#>     rowWeightedMads, rowWeightedMeans, rowWeightedMedians,
#>     rowWeightedSds, rowWeightedVars
#> Loading required package: GenomicRanges
#> Loading required package: stats4
#> Loading required package: BiocGenerics
#> 
#> Attaching package: ‘BiocGenerics’
#> The following objects are masked from ‘package:stats’:
#> 
#>     IQR, mad, sd, var, xtabs
#> The following objects are masked from ‘package:base’:
#> 
#>     Filter, Find, Map, Position, Reduce, anyDuplicated, aperm, append,
#>     as.data.frame, basename, cbind, colnames, dirname, do.call,
#>     duplicated, eval, evalq, get, grep, grepl, intersect, is.unsorted,
#>     lapply, mapply, match, mget, order, paste, pmax, pmax.int, pmin,
#>     pmin.int, rank, rbind, rownames, sapply, setdiff, table, tapply,
#>     union, unique, unsplit, which.max, which.min
#> Loading required package: S4Vectors
#> 
#> Attaching package: ‘S4Vectors’
#> The following object is masked from ‘package:tidyr’:
#> 
#>     expand
#> The following object is masked from ‘package:utils’:
#> 
#>     findMatches
#> The following objects are masked from ‘package:base’:
#> 
#>     I, expand.grid, unname
#> Loading required package: IRanges
#> Loading required package: GenomeInfoDb
#> Loading required package: Biobase
#> Welcome to Bioconductor
#> 
#>     Vignettes contain introductory material; view with
#>     'browseVignettes()'. To cite Bioconductor, see
#>     'citation("Biobase")', and for packages 'citation("pkgname")'.
#> 
#> Attaching package: ‘Biobase’
#> The following object is masked from ‘package:MatrixGenerics’:
#> 
#>     rowMedians
#> The following objects are masked from ‘package:matrixStats’:
#> 
#>     anyMissing, rowMedians
#> Loading required package: SingleCellExperiment
#> Loading required package: TreeSummarizedExperiment
#> Loading required package: Biostrings
#> Loading required package: XVector
#> 
#> Attaching package: ‘Biostrings’
#> The following object is masked from ‘package:base’:
#> 
#>     strsplit
#> This is mia version 1.15.6
#> - Online documentation and vignettes: https://microbiome.github.io/mia/
#> - Online book 'Orchestrating Microbiome Analysis (OMA)': https://microbiome.github.io/OMA/docs/devel/
library(TreeSummarizedExperiment)
library(MultiAssayExperiment)

web <- randomWeb(n_samples = 100)
mae <- as(web, "MultiAssayExperiment")

# Perform anansi analysis
out <- getAnansi(mae,
  tableY = "y", tableX = "x",
  formula = ~ cat_ab
)
#> Fitting least-squares for following model:
#> ~ x + cat_ab + x:cat_ab 
#> Running correlations for the following groups:
#>  b, a

# View subset of results
head(out, 5)
#>   feature_Y feature_X All_r.values All_t.values All_p.values  b_r.values
#> 1       y_1       x_1   0.04003613    0.3966555    0.6924839  0.02883378
#> 2       y_3       x_1  -0.09955565    0.9904713    0.3243830 -0.05567436
#> 3       y_6       x_1   0.01014953    0.1004804    0.9201682  0.12064366
#> 4      y_11       x_1  -0.01201618    0.1189627    0.9055484 -0.01744145
#> 5      y_12       x_1  -0.02639171    0.2613556    0.7943667 -0.20687414
#>   b_t.values b_p.values   a_r.values a_t.values a_p.values full_r.squared
#> 1  0.2039704  0.8392045  0.054669592 0.37134255  0.7120873   0.0016615511
#> 2  0.3942887  0.6950446 -0.135488817 0.92748230  0.3585173   0.0121164224
#> 3  0.8593564  0.3942454 -0.084344044 0.57409482  0.5687004   0.0103569349
#> 4  0.1233484  0.9023257 -0.005851666 0.03968861  0.9685131   0.0001806187
#> 5  1.4951652  0.1411566  0.125499264 0.85796070  0.3953625   0.0273601709
#>   full_f.values full_p.values disjointed_cat_ab_r.squared
#> 1   0.079887190     0.9232818                5.491606e-05
#> 2   0.588721472     0.5570277                2.233530e-03
#> 3   0.502335531     0.6066974                1.025705e-02
#> 4   0.008671265     0.9913670                4.166482e-05
#> 5   1.350230748     0.2640574                2.668074e-02
#>   disjointed_cat_ab_f.values disjointed_cat_ab_p.values
#> 1                0.005272232                  0.9422674
#> 2                0.214898904                  0.6440038
#> 3                0.994881185                  0.3210598
#> 4                0.003999989                  0.9497025
#> 5                2.631563499                  0.1080382
#>   emergent_cat_ab_r.squared emergent_cat_ab_f.values emergent_cat_ab_p.values
#> 1               0.006250521                0.5975344                0.4414394
#> 2               0.006250521                0.5975344                0.4414394
#> 3               0.006250521                0.5975344                0.4414394
#> 4               0.006250521                0.5975344                0.4414394
#> 5               0.006250521                0.5975344                0.4414394
#>   All_q.values b_q.values a_q.values full_q.values disjointed_cat_ab_q.values
#> 1    0.8473245  0.9462723  0.9362528     0.9634245                  0.9732616
#> 2    0.7773733  0.9040501  0.8364027     0.8765382                  0.9732616
#> 3    0.9397463  0.8277679  0.9099207     0.8765382                  0.9732616
#> 4    0.9397463  0.9624808  0.9891198     0.9913670                  0.9732616
#> 5    0.8473245  0.8211322  0.8364027     0.8765382                  0.9732616
#>   emergent_cat_ab_q.values
#> 1                0.6054026
#> 2                0.6054026
#> 3                0.6054026
#> 4                0.6054026
#> 5                0.6054026