anansi wrapper for the MultiAssayExperiment class

getAnansi provides a wrapper to execute the anansi pipeline on a MultiAssayExperiment object. It applies the functions weaveWeb and anansi in the given order.

Usage

getAnansi(x, ...)

# S4 method for class 'MultiAssayExperiment'
getAnansi(
  x,
  experimentY = 1,
  experimentX = 2,
  assay.typeY = "counts",
  assay.typeX = "counts",
  ...
)

Arguments

x

a MultiAssayExperiment.

...

additional parameters that can be passed to AnansiWeb or anansi.

experimentY, experimentX

Character scalar or numeric scalar. Selects experiment corresponding to tableY and tableX from experiments(x) of MultiAssayExperiment object by name or index, name is recommended. (Default slots: Y = 1, X = 2).

assay.typeY, assay.typeX

Character scalar. Specifies the name of the assay in experiments Y and X to be used. (Default: "counts")

Value

If return.format is "table" (default), a wide format data.frame intended to be compatible with `ggplot2`, or specialized plotting functions (See plotAnansi). If return.format is "list", a list with aforementioned table, as well as input and additional information. If return.format is "raw", a list of raw output (used for testing purposes).

Details

This wrapper of anansi allows to perform a complete anansi analysis directly on objects of class MultiAssayExperiment . First, the assays specified by assay.typeY and assay.typeX are passed to AnansiWeb to build an AnansiWeb object. Next, this object is fed to the main anansi function to compute interactions between the two assays.

See also

Examples

# Import libraries
library(mia)
#> Loading required package: MultiAssayExperiment
#> Loading required package: SummarizedExperiment
#> Loading required package: MatrixGenerics
#> Loading required package: matrixStats
#> 
#> Attaching package: ‘MatrixGenerics’
#> The following objects are masked from ‘package:matrixStats’:
#> 
#>     colAlls, colAnyNAs, colAnys, colAvgsPerRowSet, colCollapse,
#>     colCounts, colCummaxs, colCummins, colCumprods, colCumsums,
#>     colDiffs, colIQRDiffs, colIQRs, colLogSumExps, colMadDiffs,
#>     colMads, colMaxs, colMeans2, colMedians, colMins, colOrderStats,
#>     colProds, colQuantiles, colRanges, colRanks, colSdDiffs, colSds,
#>     colSums2, colTabulates, colVarDiffs, colVars, colWeightedMads,
#>     colWeightedMeans, colWeightedMedians, colWeightedSds,
#>     colWeightedVars, rowAlls, rowAnyNAs, rowAnys, rowAvgsPerColSet,
#>     rowCollapse, rowCounts, rowCummaxs, rowCummins, rowCumprods,
#>     rowCumsums, rowDiffs, rowIQRDiffs, rowIQRs, rowLogSumExps,
#>     rowMadDiffs, rowMads, rowMaxs, rowMeans2, rowMedians, rowMins,
#>     rowOrderStats, rowProds, rowQuantiles, rowRanges, rowRanks,
#>     rowSdDiffs, rowSds, rowSums2, rowTabulates, rowVarDiffs, rowVars,
#>     rowWeightedMads, rowWeightedMeans, rowWeightedMedians,
#>     rowWeightedSds, rowWeightedVars
#> Loading required package: GenomicRanges
#> Loading required package: stats4
#> Loading required package: BiocGenerics
#> 
#> Attaching package: ‘BiocGenerics’
#> The following objects are masked from ‘package:stats’:
#> 
#>     IQR, mad, sd, var, xtabs
#> The following objects are masked from ‘package:base’:
#> 
#>     Filter, Find, Map, Position, Reduce, anyDuplicated, aperm, append,
#>     as.data.frame, basename, cbind, colnames, dirname, do.call,
#>     duplicated, eval, evalq, get, grep, grepl, intersect, is.unsorted,
#>     lapply, mapply, match, mget, order, paste, pmax, pmax.int, pmin,
#>     pmin.int, rank, rbind, rownames, sapply, saveRDS, setdiff, table,
#>     tapply, union, unique, unsplit, which.max, which.min
#> Loading required package: S4Vectors
#> 
#> Attaching package: ‘S4Vectors’
#> The following object is masked from ‘package:tidyr’:
#> 
#>     expand
#> The following object is masked from ‘package:utils’:
#> 
#>     findMatches
#> The following objects are masked from ‘package:base’:
#> 
#>     I, expand.grid, unname
#> Loading required package: IRanges
#> Loading required package: GenomeInfoDb
#> Loading required package: Biobase
#> Welcome to Bioconductor
#> 
#>     Vignettes contain introductory material; view with
#>     'browseVignettes()'. To cite Bioconductor, see
#>     'citation("Biobase")', and for packages 'citation("pkgname")'.
#> 
#> Attaching package: ‘Biobase’
#> The following object is masked from ‘package:MatrixGenerics’:
#> 
#>     rowMedians
#> The following objects are masked from ‘package:matrixStats’:
#> 
#>     anyMissing, rowMedians
#> Loading required package: SingleCellExperiment
#> Loading required package: TreeSummarizedExperiment
#> Loading required package: Biostrings
#> Loading required package: XVector
#> 
#> Attaching package: ‘Biostrings’
#> The following object is masked from ‘package:base’:
#> 
#>     strsplit
#> This is mia version 1.14.0
#> - Online documentation and vignettes: https://microbiome.github.io/mia/
#> - Online book 'Orchestrating Microbiome Analysis (OMA)': https://microbiome.github.io/OMA/docs/devel/
library(TreeSummarizedExperiment)
library(MultiAssayExperiment)

# Load data
data("FMT_data", package = "anansi")

# Convert to (Tree)SummarizedExperiment objects
metab_se <- SummarizedExperiment(assays = SimpleList(conc = as.matrix(FMT_metab)))
KO_tse <- TreeSummarizedExperiment(assays = SimpleList(counts = as.matrix(FMT_KOs)))

# Select functions that are represented in the dictionary
keep <- row.names(KO_tse) %in% sort(unique(ec2ko$ko))
KO_tse <- KO_tse[keep, ]

# Remove features with less than 10% prevalence
KO_tse <- subsetByPrevalent(KO_tse,
  assay.type = "counts",
  prevalence = 0.1
)

# Perform a centered log-ratio transformation on the functional counts assay
KO_tse <- transformAssay(KO_tse,
  assay.type = "counts",
  method = "clr",
  pseudocount = TRUE
)
#> A pseudocount of 0.125 was applied.

# Prepare colData
coldata <- FMT_metadata
rownames(coldata) <- coldata$Sample_ID
coldata <- coldata[match(colnames(KO_tse), rownames(coldata)), ]

# Combine experiments into MultiAssayExperiment object
mae <- MultiAssayExperiment(
  experiments = ExperimentList(cpd = metab_se, ko = KO_tse),
  colData = coldata
)

# Perform anansi analysis
out <- getAnansi(mae,
  experimentY = "cpd", experimentX = "ko",
  assay.typeY = "conc", assay.typeX = "clr",
  formula = ~Legend
)
#> Fitting least-squares for following model:
#> ~ x + Legend + x:Legend 
#> Running correlations for the following groups:
#>  Aged yFMT, Aged oFMT, Young yFMT

# Select significant interactions
out <- out[out$full_p.values < 0.05, ]

# View subset of results
head(out, 5)
#>    feature_Y feature_X All_r.values All_t.values All_p.values
#> 17    C00064    K01665   -0.5693164    4.0379296 0.0002904218
#> 35    C00064    K02500    0.1363534    0.8025659 0.4278002898
#> 36    C00064    K02501    0.1349050    0.7938818 0.4327703764
#>    Aged yFMT_r.values Aged yFMT_t.values Aged yFMT_p.values Aged oFMT_r.values
#> 17        -0.01132232         0.03580661          0.9721412         -0.7590030
#> 35        -0.30753895         1.02205702          0.3308428          0.6823789
#> 36        -0.30982380         1.03045337          0.3270757          0.6811466
#>    Aged oFMT_t.values Aged oFMT_p.values Young yFMT_r.values
#> 17           3.686417        0.004201625          -0.1842965
#> 35           2.951950        0.014486249          -0.3401768
#> 36           2.941998        0.014735054          -0.3311831
#>    Young yFMT_t.values Young yFMT_p.values full_r.squared full_f.values
#> 17           0.5929537           0.5663795      0.2718152      3.732778
#> 35           1.1439573           0.2792858      0.2694013      3.687405
#> 36           1.1099303           0.2930057      0.2670031      3.642623
#>    full_p.values disjointed_Legend_r.squared disjointed_Legend_f.values
#> 17    0.02160709                   0.0833114                   1.363245
#> 35    0.02262470                   0.2637960                   5.374789
#> 36    0.02367867                   0.2610698                   5.299618
#>    disjointed_Legend_p.values emergent_Legend_r.squared
#> 17                 0.27122359                 0.1152378
#> 35                 0.01011515                 0.1619624
#> 36                 0.01069181                 0.1648405
#>    emergent_Legend_f.values emergent_Legend_p.values All_q.values
#> 17                 1.888585               0.16942785   0.01510194
#> 35                 2.802326               0.07714733   0.68616587
#> 36                 2.861952               0.07339338   0.68616587
#>    Aged yFMT_q.values Aged oFMT_q.values Young yFMT_q.values full_q.values
#> 17          0.9809225         0.07282817           0.8621558     0.4104303
#> 35          0.4915379         0.15324457           0.7763630     0.4104303
#> 36          0.4915379         0.15324457           0.7763630     0.4104303
#>    disjointed_Legend_q.values emergent_Legend_q.values
#> 17                  0.5280090                0.5182499
#> 35                  0.2779869                0.3646965
#> 36                  0.2779869                0.3646965