weaveWeb wrapper for the MultiAssayExperiment class

weaveWeb wrapper for the MultiAssayExperiment class

Usage

# S4 method for class 'MultiAssayExperiment'
getWeb(
  x,
  experimentY = 1,
  experimentX = 2,
  assay.typeY = "counts",
  assay.typeX = "counts",
  link = NULL,
  force_new = FALSE,
  ...
)

Arguments

x

a MultiAssayExperiment.

...

additional parameters that can be passed to web.

experimentY, experimentX

Character scalar or numeric scalar. Selects experiment corresponding to tableY and tableX from experiments(x) of MultiAssayExperiment object by name or index, name is recommended. (Default slots: Y = 1, X = 2).

assay.typeY, assay.typeX

Character scalar. Specifies the name of the assay in experiments Y and X to be used. (Default: "counts")

link

One of the following:

Character scalar with value "none"

data.frame with two columns

list with two such data.frames

force_new

boolean If x already has a dictionary Matrix in metadata, ignore it and generate a new object anyway? (Default: FALSE).

Value

an anansiWeb object, with sparse binary biadjacency matrix with features from y as rows and features from x as columns in dictionary slot. If x already contains a dictionary in metadata, use that one, unless force_new = TRUE.

Details

This wrapper of weaveWeb allows to generate an anansiWeb S4 object directly from objects of class MultiAssayExperiment . First, the assays specified by assay.typeY and assay.typeX are passed to web to build an anansiWeb object.

Examples

# Import libraries
library(mia)
library(TreeSummarizedExperiment)
library(MultiAssayExperiment)

# Load data
data("FMT_data", package = "anansi")

# Convert to (Tree)SummarizedExperiment objects
metab_se <- SummarizedExperiment(assays = SimpleList(conc = as.matrix(FMT_metab)))
KO_tse <- TreeSummarizedExperiment(assays = SimpleList(counts = as.matrix(FMT_KOs)))

# Remove features with less than 10% prevalence
KO_tse <- subsetByPrevalent(KO_tse,
  assay.type = "counts",
  prevalence = 0.1
)

# Perform a centered log-ratio transformation on the functional counts assay
KO_tse <- transformAssay(KO_tse,
  assay.type = "counts",
  method = "clr",
  pseudocount = TRUE
)
#> A pseudocount of 0.107142857142857 was applied.

# Prepare colData
coldata <- FMT_metadata
rownames(coldata) <- coldata$Sample_ID
coldata <- coldata[match(colnames(KO_tse), rownames(coldata)), ]

# Combine experiments into MultiAssayExperiment object
mae <- MultiAssayExperiment(
  experiments = ExperimentList(cpd = metab_se, ko = KO_tse),
  colData = coldata
)

# Perform anansi analysis
outWeb <- getWeb(mae,
  experimentY = "cpd", experimentX = "ko",
  assay.typeY = "conc", assay.typeX = "clr"
)