Generate a random AnansiWeb or MultiFactor

Randomly generate a valid AnansiWeb or MultiFactor object.

Usage

randomWeb(
  n_samples = 10,
  n_reps = 1L,
  n_features_x = 8,
  n_features_y = 12,
  sparseness = 0.5,
  tableY = NULL,
  tableX = NULL,
  dictionary = NULL
)

randomMultiFactor(n_types = 6, n_features = 100, sparseness = 0.5)

krebsDemoWeb(n_samples = 100, n_reps = 4L)

Arguments

n_samples, n_reps

Numeric scalar Number of samples and repeated measures of those samples to be generated. Ignored if tableY and tableX are provided. (defaults: 10 samples without repeats)

n_features_y, n_features_x

Numeric scalar Number of features to be generated. Ignored if tableY and tableX are provided.

sparseness

Numeric scalar, proportion: How rare are connections

tableY, tableX

A table containing features of interest. Rows should be samples and columns should be features. Y and X refer to the position of the features in a formula: Y ~ X.

dictionary

A binary adjacency matrix of class Matrix, or coercible to Matrix.

n_types

Numeric scalar, number of types of features to generate

n_features

Numeric scalar, number of features per type

Value

a randomly generated object of the specified class.

See also

Examples

# Make a random AnansiWeb object
randomWeb()
#>                       sample_id repeated group_ab subtype    score_a    score_b
#> anansi_ID_sample_1_1   sample_1    rep_1        a       z  0.8001046  0.3157294
#> anansi_ID_sample_2_1   sample_2    rep_1        b       x  1.3905638 -0.5386136
#> anansi_ID_sample_3_1   sample_3    rep_1        a       y -0.6182383  1.2855862
#> anansi_ID_sample_4_1   sample_4    rep_1        b       y -0.2047284 -0.6355094
#> anansi_ID_sample_5_1   sample_5    rep_1        a       x  0.4477842 -0.2236554
#> anansi_ID_sample_6_1   sample_6    rep_1        a       y -0.8002017  1.6115209
#> anansi_ID_sample_7_1   sample_7    rep_1        a       y -0.1493493  0.8658650
#> anansi_ID_sample_8_1   sample_8    rep_1        b       z  1.7756506 -0.9232705
#> anansi_ID_sample_9_1   sample_9    rep_1        a       y -0.1417170 -1.5596819
#> anansi_ID_sample_10_1 sample_10    rep_1        a       z  1.3243586  0.4326524
#>                           score_c
#> anansi_ID_sample_1_1   0.31906949
#> anansi_ID_sample_2_1   0.89061572
#> anansi_ID_sample_3_1  -0.14285348
#> anansi_ID_sample_4_1  -0.28776497
#> anansi_ID_sample_5_1   0.20490668
#> anansi_ID_sample_6_1  -0.09795202
#> anansi_ID_sample_7_1   1.96248885
#> anansi_ID_sample_8_1  -1.49375735
#> anansi_ID_sample_9_1   0.37443364
#> anansi_ID_sample_10_1  0.33399321
#> anansi::AnansiWeb S7_object with 10 observations:
#>     tableY: y (12 features)
#>     tableX: x (8 features)
#> Use @ to access: tableX, tableY, dictionary, metadata.
krebsDemoWeb()
#>                         sample_id repeated group_ab subtype      score_a
#> anansi_ID_sample_1_1     sample_1    rep_1        b       y  1.385378955
#> anansi_ID_sample_1_2     sample_1    rep_2        b       y  0.652462045
#> anansi_ID_sample_1_3     sample_1    rep_3        b       y -0.461044084
#> anansi_ID_sample_1_4     sample_1    rep_4        b       y  1.030935989
#> anansi_ID_sample_2_1     sample_2    rep_1        a       z -0.750069512
#> anansi_ID_sample_2_2     sample_2    rep_2        a       z  0.276435764
#> anansi_ID_sample_2_3     sample_2    rep_3        a       z  0.393342785
#> anansi_ID_sample_2_4     sample_2    rep_4        a       z  1.230616240
#> anansi_ID_sample_3_1     sample_3    rep_1        a       x  0.339242130
#> anansi_ID_sample_3_2     sample_3    rep_2        a       x  1.076904414
#> anansi_ID_sample_3_3     sample_3    rep_3        a       x -1.171719711
#> anansi_ID_sample_3_4     sample_3    rep_4        a       x -1.118750535
#> anansi_ID_sample_4_1     sample_4    rep_1        a       z  1.070001873
#> anansi_ID_sample_4_2     sample_4    rep_2        a       z -1.884003713
#> anansi_ID_sample_4_3     sample_4    rep_3        a       z -0.830953876
#> anansi_ID_sample_4_4     sample_4    rep_4        a       z  0.210906410
#> anansi_ID_sample_5_1     sample_5    rep_1        a       y -1.190189278
#> anansi_ID_sample_5_2     sample_5    rep_2        a       y  0.988604262
#> anansi_ID_sample_5_3     sample_5    rep_3        a       y -0.815854999
#> anansi_ID_sample_5_4     sample_5    rep_4        a       y -1.303932492
#> anansi_ID_sample_6_1     sample_6    rep_1        b       y -0.274133653
#> anansi_ID_sample_6_2     sample_6    rep_2        b       y -0.400166691
#> anansi_ID_sample_6_3     sample_6    rep_3        b       y  0.720642291
#> anansi_ID_sample_6_4     sample_6    rep_4        b       y -0.443131800
#> anansi_ID_sample_7_1     sample_7    rep_1        b       x  0.156717952
#> anansi_ID_sample_7_2     sample_7    rep_2        b       x -2.120529788
#> anansi_ID_sample_7_3     sample_7    rep_3        b       x  0.245580472
#> anansi_ID_sample_7_4     sample_7    rep_4        b       x  2.151837028
#> anansi_ID_sample_8_1     sample_8    rep_1        b       x  1.477719793
#> anansi_ID_sample_8_2     sample_8    rep_2        b       x  1.042294664
#> anansi_ID_sample_8_3     sample_8    rep_3        b       x -0.610958326
#> anansi_ID_sample_8_4     sample_8    rep_4        b       x -0.015217060
#> anansi_ID_sample_9_1     sample_9    rep_1        b       y -1.851830483
#> anansi_ID_sample_9_2     sample_9    rep_2        b       y -0.832774109
#> anansi_ID_sample_9_3     sample_9    rep_3        b       y -1.393115933
#> anansi_ID_sample_9_4     sample_9    rep_4        b       y  0.088582035
#> anansi_ID_sample_10_1   sample_10    rep_1        a       y -0.785499290
#> anansi_ID_sample_10_2   sample_10    rep_2        a       y -0.184082406
#> anansi_ID_sample_10_3   sample_10    rep_3        a       y  2.275370512
#> anansi_ID_sample_10_4   sample_10    rep_4        a       y -0.789559324
#> anansi_ID_sample_11_1   sample_11    rep_1        a       z -1.558399097
#> anansi_ID_sample_11_2   sample_11    rep_2        a       z -0.256191813
#> anansi_ID_sample_11_3   sample_11    rep_3        a       z  1.612869148
#> anansi_ID_sample_11_4   sample_11    rep_4        a       z -0.694988239
#> anansi_ID_sample_12_1   sample_12    rep_1        b       y -0.483748677
#> anansi_ID_sample_12_2   sample_12    rep_2        b       y -0.324011498
#> anansi_ID_sample_12_3   sample_12    rep_3        b       y -1.171198770
#> anansi_ID_sample_12_4   sample_12    rep_4        b       y -1.701421547
#> anansi_ID_sample_13_1   sample_13    rep_1        a       z -0.836523723
#> anansi_ID_sample_13_2   sample_13    rep_2        a       z -0.214193060
#> anansi_ID_sample_13_3   sample_13    rep_3        a       z  0.150531519
#> anansi_ID_sample_13_4   sample_13    rep_4        a       z -0.364691456
#> anansi_ID_sample_14_1   sample_14    rep_1        a       y  0.477596646
#> anansi_ID_sample_14_2   sample_14    rep_2        a       y -0.393687228
#> anansi_ID_sample_14_3   sample_14    rep_3        a       y  0.471319883
#> anansi_ID_sample_14_4   sample_14    rep_4        a       y -0.780122108
#> anansi_ID_sample_15_1   sample_15    rep_1        a       x -0.193622740
#> anansi_ID_sample_15_2   sample_15    rep_2        a       x -1.237170560
#> anansi_ID_sample_15_3   sample_15    rep_3        a       x -2.248786029
#> anansi_ID_sample_15_4   sample_15    rep_4        a       x -0.327179442
#> anansi_ID_sample_16_1   sample_16    rep_1        b       y  0.972333693
#> anansi_ID_sample_16_2   sample_16    rep_2        b       y -0.120617700
#> anansi_ID_sample_16_3   sample_16    rep_3        b       y  1.882270101
#> anansi_ID_sample_16_4   sample_16    rep_4        b       y  0.560476483
#> anansi_ID_sample_17_1   sample_17    rep_1        a       y -2.429148121
#> anansi_ID_sample_17_2   sample_17    rep_2        a       y -0.668798310
#> anansi_ID_sample_17_3   sample_17    rep_3        a       y -2.158352002
#> anansi_ID_sample_17_4   sample_17    rep_4        a       y -0.169416348
#> anansi_ID_sample_18_1   sample_18    rep_1        b       z  0.001528965
#> anansi_ID_sample_18_2   sample_18    rep_2        b       z  0.819025340
#> anansi_ID_sample_18_3   sample_18    rep_3        b       z -0.770021636
#> anansi_ID_sample_18_4   sample_18    rep_4        b       z -1.041612479
#> anansi_ID_sample_19_1   sample_19    rep_1        a       z -0.961786822
#> anansi_ID_sample_19_2   sample_19    rep_2        a       z  0.578339827
#> anansi_ID_sample_19_3   sample_19    rep_3        a       z -0.192353394
#> anansi_ID_sample_19_4   sample_19    rep_4        a       z  0.295748762
#> anansi_ID_sample_20_1   sample_20    rep_1        b       y -1.391502485
#> anansi_ID_sample_20_2   sample_20    rep_2        b       y -0.456859603
#> anansi_ID_sample_20_3   sample_20    rep_3        b       y  1.188637774
#> anansi_ID_sample_20_4   sample_20    rep_4        b       y  0.135813285
#> anansi_ID_sample_21_1   sample_21    rep_1        b       y -0.039440547
#> anansi_ID_sample_21_2   sample_21    rep_2        b       y -1.830625148
#> anansi_ID_sample_21_3   sample_21    rep_3        b       y -0.040319706
#> anansi_ID_sample_21_4   sample_21    rep_4        b       y  0.102643506
#> anansi_ID_sample_22_1   sample_22    rep_1        a       y  0.570353758
#> anansi_ID_sample_22_2   sample_22    rep_2        a       y -0.122340037
#> anansi_ID_sample_22_3   sample_22    rep_3        a       y  0.526152689
#> anansi_ID_sample_22_4   sample_22    rep_4        a       y  1.226000236
#> anansi_ID_sample_23_1   sample_23    rep_1        a       y -0.440374436
#> anansi_ID_sample_23_2   sample_23    rep_2        a       y -2.021389808
#> anansi_ID_sample_23_3   sample_23    rep_3        a       y  0.820660916
#> anansi_ID_sample_23_4   sample_23    rep_4        a       y -0.441832309
#> anansi_ID_sample_24_1   sample_24    rep_1        b       z  1.420463195
#> anansi_ID_sample_24_2   sample_24    rep_2        b       z -0.698350575
#> anansi_ID_sample_24_3   sample_24    rep_3        b       z  0.129206549
#> anansi_ID_sample_24_4   sample_24    rep_4        b       z -1.469734086
#> anansi_ID_sample_25_1   sample_25    rep_1        b       y  0.235709661
#> anansi_ID_sample_25_2   sample_25    rep_2        b       y  2.279130730
#> anansi_ID_sample_25_3   sample_25    rep_3        b       y  0.224881954
#> anansi_ID_sample_25_4   sample_25    rep_4        b       y  0.129326660
#> anansi_ID_sample_26_1   sample_26    rep_1        b       z  1.385378955
#> anansi_ID_sample_26_2   sample_26    rep_2        b       z  0.652462045
#> anansi_ID_sample_26_3   sample_26    rep_3        b       z -0.461044084
#> anansi_ID_sample_26_4   sample_26    rep_4        b       z  1.030935989
#> anansi_ID_sample_27_1   sample_27    rep_1        a       z -0.750069512
#> anansi_ID_sample_27_2   sample_27    rep_2        a       z  0.276435764
#> anansi_ID_sample_27_3   sample_27    rep_3        a       z  0.393342785
#> anansi_ID_sample_27_4   sample_27    rep_4        a       z  1.230616240
#> anansi_ID_sample_28_1   sample_28    rep_1        b       y  0.339242130
#> anansi_ID_sample_28_2   sample_28    rep_2        b       y  1.076904414
#> anansi_ID_sample_28_3   sample_28    rep_3        b       y -1.171719711
#> anansi_ID_sample_28_4   sample_28    rep_4        b       y -1.118750535
#> anansi_ID_sample_29_1   sample_29    rep_1        a       y  1.070001873
#> anansi_ID_sample_29_2   sample_29    rep_2        a       y -1.884003713
#> anansi_ID_sample_29_3   sample_29    rep_3        a       y -0.830953876
#> anansi_ID_sample_29_4   sample_29    rep_4        a       y  0.210906410
#> anansi_ID_sample_30_1   sample_30    rep_1        b       x -1.190189278
#> anansi_ID_sample_30_2   sample_30    rep_2        b       x  0.988604262
#> anansi_ID_sample_30_3   sample_30    rep_3        b       x -0.815854999
#> anansi_ID_sample_30_4   sample_30    rep_4        b       x -1.303932492
#> anansi_ID_sample_31_1   sample_31    rep_1        b       z -0.274133653
#> anansi_ID_sample_31_2   sample_31    rep_2        b       z -0.400166691
#> anansi_ID_sample_31_3   sample_31    rep_3        b       z  0.720642291
#> anansi_ID_sample_31_4   sample_31    rep_4        b       z -0.443131800
#> anansi_ID_sample_32_1   sample_32    rep_1        b       x  0.156717952
#> anansi_ID_sample_32_2   sample_32    rep_2        b       x -2.120529788
#> anansi_ID_sample_32_3   sample_32    rep_3        b       x  0.245580472
#> anansi_ID_sample_32_4   sample_32    rep_4        b       x  2.151837028
#> anansi_ID_sample_33_1   sample_33    rep_1        b       y  1.477719793
#> anansi_ID_sample_33_2   sample_33    rep_2        b       y  1.042294664
#> anansi_ID_sample_33_3   sample_33    rep_3        b       y -0.610958326
#> anansi_ID_sample_33_4   sample_33    rep_4        b       y -0.015217060
#> anansi_ID_sample_34_1   sample_34    rep_1        b       x -1.851830483
#> anansi_ID_sample_34_2   sample_34    rep_2        b       x -0.832774109
#> anansi_ID_sample_34_3   sample_34    rep_3        b       x -1.393115933
#> anansi_ID_sample_34_4   sample_34    rep_4        b       x  0.088582035
#> anansi_ID_sample_35_1   sample_35    rep_1        b       z -0.785499290
#> anansi_ID_sample_35_2   sample_35    rep_2        b       z -0.184082406
#> anansi_ID_sample_35_3   sample_35    rep_3        b       z  2.275370512
#> anansi_ID_sample_35_4   sample_35    rep_4        b       z -0.789559324
#> anansi_ID_sample_36_1   sample_36    rep_1        a       z -1.558399097
#> anansi_ID_sample_36_2   sample_36    rep_2        a       z -0.256191813
#> anansi_ID_sample_36_3   sample_36    rep_3        a       z  1.612869148
#> anansi_ID_sample_36_4   sample_36    rep_4        a       z -0.694988239
#> anansi_ID_sample_37_1   sample_37    rep_1        a       x -0.483748677
#> anansi_ID_sample_37_2   sample_37    rep_2        a       x -0.324011498
#> anansi_ID_sample_37_3   sample_37    rep_3        a       x -1.171198770
#> anansi_ID_sample_37_4   sample_37    rep_4        a       x -1.701421547
#> anansi_ID_sample_38_1   sample_38    rep_1        a       z -0.836523723
#> anansi_ID_sample_38_2   sample_38    rep_2        a       z -0.214193060
#> anansi_ID_sample_38_3   sample_38    rep_3        a       z  0.150531519
#> anansi_ID_sample_38_4   sample_38    rep_4        a       z -0.364691456
#> anansi_ID_sample_39_1   sample_39    rep_1        b       x  0.477596646
#> anansi_ID_sample_39_2   sample_39    rep_2        b       x -0.393687228
#> anansi_ID_sample_39_3   sample_39    rep_3        b       x  0.471319883
#> anansi_ID_sample_39_4   sample_39    rep_4        b       x -0.780122108
#> anansi_ID_sample_40_1   sample_40    rep_1        b       x -0.193622740
#> anansi_ID_sample_40_2   sample_40    rep_2        b       x -1.237170560
#> anansi_ID_sample_40_3   sample_40    rep_3        b       x -2.248786029
#> anansi_ID_sample_40_4   sample_40    rep_4        b       x -0.327179442
#> anansi_ID_sample_41_1   sample_41    rep_1        a       y  0.972333693
#> anansi_ID_sample_41_2   sample_41    rep_2        a       y -0.120617700
#> anansi_ID_sample_41_3   sample_41    rep_3        a       y  1.882270101
#> anansi_ID_sample_41_4   sample_41    rep_4        a       y  0.560476483
#> anansi_ID_sample_42_1   sample_42    rep_1        a       y -2.429148121
#> anansi_ID_sample_42_2   sample_42    rep_2        a       y -0.668798310
#> anansi_ID_sample_42_3   sample_42    rep_3        a       y -2.158352002
#> anansi_ID_sample_42_4   sample_42    rep_4        a       y -0.169416348
#> anansi_ID_sample_43_1   sample_43    rep_1        a       y  0.001528965
#> anansi_ID_sample_43_2   sample_43    rep_2        a       y  0.819025340
#> anansi_ID_sample_43_3   sample_43    rep_3        a       y -0.770021636
#> anansi_ID_sample_43_4   sample_43    rep_4        a       y -1.041612479
#> anansi_ID_sample_44_1   sample_44    rep_1        a       z -0.961786822
#> anansi_ID_sample_44_2   sample_44    rep_2        a       z  0.578339827
#> anansi_ID_sample_44_3   sample_44    rep_3        a       z -0.192353394
#> anansi_ID_sample_44_4   sample_44    rep_4        a       z  0.295748762
#> anansi_ID_sample_45_1   sample_45    rep_1        a       x -1.391502485
#> anansi_ID_sample_45_2   sample_45    rep_2        a       x -0.456859603
#> anansi_ID_sample_45_3   sample_45    rep_3        a       x  1.188637774
#> anansi_ID_sample_45_4   sample_45    rep_4        a       x  0.135813285
#> anansi_ID_sample_46_1   sample_46    rep_1        b       y -0.039440547
#> anansi_ID_sample_46_2   sample_46    rep_2        b       y -1.830625148
#> anansi_ID_sample_46_3   sample_46    rep_3        b       y -0.040319706
#> anansi_ID_sample_46_4   sample_46    rep_4        b       y  0.102643506
#> anansi_ID_sample_47_1   sample_47    rep_1        b       y  0.570353758
#> anansi_ID_sample_47_2   sample_47    rep_2        b       y -0.122340037
#> anansi_ID_sample_47_3   sample_47    rep_3        b       y  0.526152689
#> anansi_ID_sample_47_4   sample_47    rep_4        b       y  1.226000236
#> anansi_ID_sample_48_1   sample_48    rep_1        b       x -0.440374436
#> anansi_ID_sample_48_2   sample_48    rep_2        b       x -2.021389808
#> anansi_ID_sample_48_3   sample_48    rep_3        b       x  0.820660916
#> anansi_ID_sample_48_4   sample_48    rep_4        b       x -0.441832309
#> anansi_ID_sample_49_1   sample_49    rep_1        b       y  1.420463195
#> anansi_ID_sample_49_2   sample_49    rep_2        b       y -0.698350575
#> anansi_ID_sample_49_3   sample_49    rep_3        b       y  0.129206549
#> anansi_ID_sample_49_4   sample_49    rep_4        b       y -1.469734086
#> anansi_ID_sample_50_1   sample_50    rep_1        a       z  0.235709661
#> anansi_ID_sample_50_2   sample_50    rep_2        a       z  2.279130730
#> anansi_ID_sample_50_3   sample_50    rep_3        a       z  0.224881954
#> anansi_ID_sample_50_4   sample_50    rep_4        a       z  0.129326660
#> anansi_ID_sample_51_1   sample_51    rep_1        b       z  1.385378955
#> anansi_ID_sample_51_2   sample_51    rep_2        b       z  0.652462045
#> anansi_ID_sample_51_3   sample_51    rep_3        b       z -0.461044084
#> anansi_ID_sample_51_4   sample_51    rep_4        b       z  1.030935989
#> anansi_ID_sample_52_1   sample_52    rep_1        b       x -0.750069512
#> anansi_ID_sample_52_2   sample_52    rep_2        b       x  0.276435764
#> anansi_ID_sample_52_3   sample_52    rep_3        b       x  0.393342785
#> anansi_ID_sample_52_4   sample_52    rep_4        b       x  1.230616240
#> anansi_ID_sample_53_1   sample_53    rep_1        a       y  0.339242130
#> anansi_ID_sample_53_2   sample_53    rep_2        a       y  1.076904414
#> anansi_ID_sample_53_3   sample_53    rep_3        a       y -1.171719711
#> anansi_ID_sample_53_4   sample_53    rep_4        a       y -1.118750535
#> anansi_ID_sample_54_1   sample_54    rep_1        a       z  1.070001873
#> anansi_ID_sample_54_2   sample_54    rep_2        a       z -1.884003713
#> anansi_ID_sample_54_3   sample_54    rep_3        a       z -0.830953876
#> anansi_ID_sample_54_4   sample_54    rep_4        a       z  0.210906410
#> anansi_ID_sample_55_1   sample_55    rep_1        a       z -1.190189278
#> anansi_ID_sample_55_2   sample_55    rep_2        a       z  0.988604262
#> anansi_ID_sample_55_3   sample_55    rep_3        a       z -0.815854999
#> anansi_ID_sample_55_4   sample_55    rep_4        a       z -1.303932492
#> anansi_ID_sample_56_1   sample_56    rep_1        a       y -0.274133653
#> anansi_ID_sample_56_2   sample_56    rep_2        a       y -0.400166691
#> anansi_ID_sample_56_3   sample_56    rep_3        a       y  0.720642291
#> anansi_ID_sample_56_4   sample_56    rep_4        a       y -0.443131800
#> anansi_ID_sample_57_1   sample_57    rep_1        b       y  0.156717952
#> anansi_ID_sample_57_2   sample_57    rep_2        b       y -2.120529788
#> anansi_ID_sample_57_3   sample_57    rep_3        b       y  0.245580472
#> anansi_ID_sample_57_4   sample_57    rep_4        b       y  2.151837028
#> anansi_ID_sample_58_1   sample_58    rep_1        b       x  1.477719793
#> anansi_ID_sample_58_2   sample_58    rep_2        b       x  1.042294664
#> anansi_ID_sample_58_3   sample_58    rep_3        b       x -0.610958326
#> anansi_ID_sample_58_4   sample_58    rep_4        b       x -0.015217060
#> anansi_ID_sample_59_1   sample_59    rep_1        b       y -1.851830483
#> anansi_ID_sample_59_2   sample_59    rep_2        b       y -0.832774109
#> anansi_ID_sample_59_3   sample_59    rep_3        b       y -1.393115933
#> anansi_ID_sample_59_4   sample_59    rep_4        b       y  0.088582035
#> anansi_ID_sample_60_1   sample_60    rep_1        a       x -0.785499290
#> anansi_ID_sample_60_2   sample_60    rep_2        a       x -0.184082406
#> anansi_ID_sample_60_3   sample_60    rep_3        a       x  2.275370512
#> anansi_ID_sample_60_4   sample_60    rep_4        a       x -0.789559324
#> anansi_ID_sample_61_1   sample_61    rep_1        a       x -1.558399097
#> anansi_ID_sample_61_2   sample_61    rep_2        a       x -0.256191813
#> anansi_ID_sample_61_3   sample_61    rep_3        a       x  1.612869148
#> anansi_ID_sample_61_4   sample_61    rep_4        a       x -0.694988239
#> anansi_ID_sample_62_1   sample_62    rep_1        b       x -0.483748677
#> anansi_ID_sample_62_2   sample_62    rep_2        b       x -0.324011498
#> anansi_ID_sample_62_3   sample_62    rep_3        b       x -1.171198770
#> anansi_ID_sample_62_4   sample_62    rep_4        b       x -1.701421547
#> anansi_ID_sample_63_1   sample_63    rep_1        b       z -0.836523723
#> anansi_ID_sample_63_2   sample_63    rep_2        b       z -0.214193060
#> anansi_ID_sample_63_3   sample_63    rep_3        b       z  0.150531519
#> anansi_ID_sample_63_4   sample_63    rep_4        b       z -0.364691456
#> anansi_ID_sample_64_1   sample_64    rep_1        b       z  0.477596646
#> anansi_ID_sample_64_2   sample_64    rep_2        b       z -0.393687228
#> anansi_ID_sample_64_3   sample_64    rep_3        b       z  0.471319883
#> anansi_ID_sample_64_4   sample_64    rep_4        b       z -0.780122108
#> anansi_ID_sample_65_1   sample_65    rep_1        a       x -0.193622740
#> anansi_ID_sample_65_2   sample_65    rep_2        a       x -1.237170560
#> anansi_ID_sample_65_3   sample_65    rep_3        a       x -2.248786029
#> anansi_ID_sample_65_4   sample_65    rep_4        a       x -0.327179442
#> anansi_ID_sample_66_1   sample_66    rep_1        b       y  0.972333693
#> anansi_ID_sample_66_2   sample_66    rep_2        b       y -0.120617700
#> anansi_ID_sample_66_3   sample_66    rep_3        b       y  1.882270101
#> anansi_ID_sample_66_4   sample_66    rep_4        b       y  0.560476483
#> anansi_ID_sample_67_1   sample_67    rep_1        b       x -2.429148121
#> anansi_ID_sample_67_2   sample_67    rep_2        b       x -0.668798310
#> anansi_ID_sample_67_3   sample_67    rep_3        b       x -2.158352002
#> anansi_ID_sample_67_4   sample_67    rep_4        b       x -0.169416348
#> anansi_ID_sample_68_1   sample_68    rep_1        a       y  0.001528965
#> anansi_ID_sample_68_2   sample_68    rep_2        a       y  0.819025340
#> anansi_ID_sample_68_3   sample_68    rep_3        a       y -0.770021636
#> anansi_ID_sample_68_4   sample_68    rep_4        a       y -1.041612479
#> anansi_ID_sample_69_1   sample_69    rep_1        b       z -0.961786822
#> anansi_ID_sample_69_2   sample_69    rep_2        b       z  0.578339827
#> anansi_ID_sample_69_3   sample_69    rep_3        b       z -0.192353394
#> anansi_ID_sample_69_4   sample_69    rep_4        b       z  0.295748762
#> anansi_ID_sample_70_1   sample_70    rep_1        b       y -1.391502485
#> anansi_ID_sample_70_2   sample_70    rep_2        b       y -0.456859603
#> anansi_ID_sample_70_3   sample_70    rep_3        b       y  1.188637774
#> anansi_ID_sample_70_4   sample_70    rep_4        b       y  0.135813285
#> anansi_ID_sample_71_1   sample_71    rep_1        a       y -0.039440547
#> anansi_ID_sample_71_2   sample_71    rep_2        a       y -1.830625148
#> anansi_ID_sample_71_3   sample_71    rep_3        a       y -0.040319706
#> anansi_ID_sample_71_4   sample_71    rep_4        a       y  0.102643506
#> anansi_ID_sample_72_1   sample_72    rep_1        b       z  0.570353758
#> anansi_ID_sample_72_2   sample_72    rep_2        b       z -0.122340037
#> anansi_ID_sample_72_3   sample_72    rep_3        b       z  0.526152689
#> anansi_ID_sample_72_4   sample_72    rep_4        b       z  1.226000236
#> anansi_ID_sample_73_1   sample_73    rep_1        b       z -0.440374436
#> anansi_ID_sample_73_2   sample_73    rep_2        b       z -2.021389808
#> anansi_ID_sample_73_3   sample_73    rep_3        b       z  0.820660916
#> anansi_ID_sample_73_4   sample_73    rep_4        b       z -0.441832309
#> anansi_ID_sample_74_1   sample_74    rep_1        a       x  1.420463195
#> anansi_ID_sample_74_2   sample_74    rep_2        a       x -0.698350575
#> anansi_ID_sample_74_3   sample_74    rep_3        a       x  0.129206549
#> anansi_ID_sample_74_4   sample_74    rep_4        a       x -1.469734086
#> anansi_ID_sample_75_1   sample_75    rep_1        b       y  0.235709661
#> anansi_ID_sample_75_2   sample_75    rep_2        b       y  2.279130730
#> anansi_ID_sample_75_3   sample_75    rep_3        b       y  0.224881954
#> anansi_ID_sample_75_4   sample_75    rep_4        b       y  0.129326660
#> anansi_ID_sample_76_1   sample_76    rep_1        a       x  1.385378955
#> anansi_ID_sample_76_2   sample_76    rep_2        a       x  0.652462045
#> anansi_ID_sample_76_3   sample_76    rep_3        a       x -0.461044084
#> anansi_ID_sample_76_4   sample_76    rep_4        a       x  1.030935989
#> anansi_ID_sample_77_1   sample_77    rep_1        b       z -0.750069512
#> anansi_ID_sample_77_2   sample_77    rep_2        b       z  0.276435764
#> anansi_ID_sample_77_3   sample_77    rep_3        b       z  0.393342785
#> anansi_ID_sample_77_4   sample_77    rep_4        b       z  1.230616240
#> anansi_ID_sample_78_1   sample_78    rep_1        b       z  0.339242130
#> anansi_ID_sample_78_2   sample_78    rep_2        b       z  1.076904414
#> anansi_ID_sample_78_3   sample_78    rep_3        b       z -1.171719711
#> anansi_ID_sample_78_4   sample_78    rep_4        b       z -1.118750535
#> anansi_ID_sample_79_1   sample_79    rep_1        a       y  1.070001873
#> anansi_ID_sample_79_2   sample_79    rep_2        a       y -1.884003713
#> anansi_ID_sample_79_3   sample_79    rep_3        a       y -0.830953876
#> anansi_ID_sample_79_4   sample_79    rep_4        a       y  0.210906410
#> anansi_ID_sample_80_1   sample_80    rep_1        b       z -1.190189278
#> anansi_ID_sample_80_2   sample_80    rep_2        b       z  0.988604262
#> anansi_ID_sample_80_3   sample_80    rep_3        b       z -0.815854999
#> anansi_ID_sample_80_4   sample_80    rep_4        b       z -1.303932492
#> anansi_ID_sample_81_1   sample_81    rep_1        b       x -0.274133653
#> anansi_ID_sample_81_2   sample_81    rep_2        b       x -0.400166691
#> anansi_ID_sample_81_3   sample_81    rep_3        b       x  0.720642291
#> anansi_ID_sample_81_4   sample_81    rep_4        b       x -0.443131800
#> anansi_ID_sample_82_1   sample_82    rep_1        b       x  0.156717952
#> anansi_ID_sample_82_2   sample_82    rep_2        b       x -2.120529788
#> anansi_ID_sample_82_3   sample_82    rep_3        b       x  0.245580472
#> anansi_ID_sample_82_4   sample_82    rep_4        b       x  2.151837028
#> anansi_ID_sample_83_1   sample_83    rep_1        a       x  1.477719793
#> anansi_ID_sample_83_2   sample_83    rep_2        a       x  1.042294664
#> anansi_ID_sample_83_3   sample_83    rep_3        a       x -0.610958326
#> anansi_ID_sample_83_4   sample_83    rep_4        a       x -0.015217060
#> anansi_ID_sample_84_1   sample_84    rep_1        a       x -1.851830483
#> anansi_ID_sample_84_2   sample_84    rep_2        a       x -0.832774109
#> anansi_ID_sample_84_3   sample_84    rep_3        a       x -1.393115933
#> anansi_ID_sample_84_4   sample_84    rep_4        a       x  0.088582035
#> anansi_ID_sample_85_1   sample_85    rep_1        b       z -0.785499290
#> anansi_ID_sample_85_2   sample_85    rep_2        b       z -0.184082406
#> anansi_ID_sample_85_3   sample_85    rep_3        b       z  2.275370512
#> anansi_ID_sample_85_4   sample_85    rep_4        b       z -0.789559324
#> anansi_ID_sample_86_1   sample_86    rep_1        a       z -1.558399097
#> anansi_ID_sample_86_2   sample_86    rep_2        a       z -0.256191813
#> anansi_ID_sample_86_3   sample_86    rep_3        a       z  1.612869148
#> anansi_ID_sample_86_4   sample_86    rep_4        a       z -0.694988239
#> anansi_ID_sample_87_1   sample_87    rep_1        b       z -0.483748677
#> anansi_ID_sample_87_2   sample_87    rep_2        b       z -0.324011498
#> anansi_ID_sample_87_3   sample_87    rep_3        b       z -1.171198770
#> anansi_ID_sample_87_4   sample_87    rep_4        b       z -1.701421547
#> anansi_ID_sample_88_1   sample_88    rep_1        b       y -0.836523723
#> anansi_ID_sample_88_2   sample_88    rep_2        b       y -0.214193060
#> anansi_ID_sample_88_3   sample_88    rep_3        b       y  0.150531519
#> anansi_ID_sample_88_4   sample_88    rep_4        b       y -0.364691456
#> anansi_ID_sample_89_1   sample_89    rep_1        b       z  0.477596646
#> anansi_ID_sample_89_2   sample_89    rep_2        b       z -0.393687228
#> anansi_ID_sample_89_3   sample_89    rep_3        b       z  0.471319883
#> anansi_ID_sample_89_4   sample_89    rep_4        b       z -0.780122108
#> anansi_ID_sample_90_1   sample_90    rep_1        a       y -0.193622740
#> anansi_ID_sample_90_2   sample_90    rep_2        a       y -1.237170560
#> anansi_ID_sample_90_3   sample_90    rep_3        a       y -2.248786029
#> anansi_ID_sample_90_4   sample_90    rep_4        a       y -0.327179442
#> anansi_ID_sample_91_1   sample_91    rep_1        b       z  0.972333693
#> anansi_ID_sample_91_2   sample_91    rep_2        b       z -0.120617700
#> anansi_ID_sample_91_3   sample_91    rep_3        b       z  1.882270101
#> anansi_ID_sample_91_4   sample_91    rep_4        b       z  0.560476483
#> anansi_ID_sample_92_1   sample_92    rep_1        b       y -2.429148121
#> anansi_ID_sample_92_2   sample_92    rep_2        b       y -0.668798310
#> anansi_ID_sample_92_3   sample_92    rep_3        b       y -2.158352002
#> anansi_ID_sample_92_4   sample_92    rep_4        b       y -0.169416348
#> anansi_ID_sample_93_1   sample_93    rep_1        a       z  0.001528965
#> anansi_ID_sample_93_2   sample_93    rep_2        a       z  0.819025340
#> anansi_ID_sample_93_3   sample_93    rep_3        a       z -0.770021636
#> anansi_ID_sample_93_4   sample_93    rep_4        a       z -1.041612479
#> anansi_ID_sample_94_1   sample_94    rep_1        a       x -0.961786822
#> anansi_ID_sample_94_2   sample_94    rep_2        a       x  0.578339827
#> anansi_ID_sample_94_3   sample_94    rep_3        a       x -0.192353394
#> anansi_ID_sample_94_4   sample_94    rep_4        a       x  0.295748762
#> anansi_ID_sample_95_1   sample_95    rep_1        b       z -1.391502485
#> anansi_ID_sample_95_2   sample_95    rep_2        b       z -0.456859603
#> anansi_ID_sample_95_3   sample_95    rep_3        b       z  1.188637774
#> anansi_ID_sample_95_4   sample_95    rep_4        b       z  0.135813285
#> anansi_ID_sample_96_1   sample_96    rep_1        a       y -0.039440547
#> anansi_ID_sample_96_2   sample_96    rep_2        a       y -1.830625148
#> anansi_ID_sample_96_3   sample_96    rep_3        a       y -0.040319706
#> anansi_ID_sample_96_4   sample_96    rep_4        a       y  0.102643506
#> anansi_ID_sample_97_1   sample_97    rep_1        a       y  0.570353758
#> anansi_ID_sample_97_2   sample_97    rep_2        a       y -0.122340037
#> anansi_ID_sample_97_3   sample_97    rep_3        a       y  0.526152689
#> anansi_ID_sample_97_4   sample_97    rep_4        a       y  1.226000236
#> anansi_ID_sample_98_1   sample_98    rep_1        b       x -0.440374436
#> anansi_ID_sample_98_2   sample_98    rep_2        b       x -2.021389808
#> anansi_ID_sample_98_3   sample_98    rep_3        b       x  0.820660916
#> anansi_ID_sample_98_4   sample_98    rep_4        b       x -0.441832309
#> anansi_ID_sample_99_1   sample_99    rep_1        a       z  1.420463195
#> anansi_ID_sample_99_2   sample_99    rep_2        a       z -0.698350575
#> anansi_ID_sample_99_3   sample_99    rep_3        a       z  0.129206549
#> anansi_ID_sample_99_4   sample_99    rep_4        a       z -1.469734086
#> anansi_ID_sample_100_1 sample_100    rep_1        b       y  0.235709661
#> anansi_ID_sample_100_2 sample_100    rep_2        b       y  2.279130730
#> anansi_ID_sample_100_3 sample_100    rep_3        b       y  0.224881954
#> anansi_ID_sample_100_4 sample_100    rep_4        b       y  0.129326660
#>                             score_b     score_c
#> anansi_ID_sample_1_1   -0.419197962  1.24660407
#> anansi_ID_sample_1_2   -0.258272124 -1.11524545
#> anansi_ID_sample_1_3    0.387013494  0.35407631
#> anansi_ID_sample_1_4   -0.514148130  0.40088126
#> anansi_ID_sample_2_1   -1.328172982  0.56616481
#> anansi_ID_sample_2_2   -0.533051253 -0.70611414
#> anansi_ID_sample_2_3   -0.296472713  0.23189206
#> anansi_ID_sample_2_4    0.233670826  0.70034619
#> anansi_ID_sample_3_1   -0.545436915  0.22092862
#> anansi_ID_sample_3_2   -0.454328382  0.94493135
#> anansi_ID_sample_3_3    1.006300835 -1.59317389
#> anansi_ID_sample_3_4    0.260808067  0.12626472
#> anansi_ID_sample_4_1   -1.116993390  0.07962046
#> anansi_ID_sample_4_2   -1.814573296  1.19316824
#> anansi_ID_sample_4_3   -0.179786139 -0.31193914
#> anansi_ID_sample_4_4    0.792354718  2.05624227
#> anansi_ID_sample_5_1    0.693294878 -0.12571184
#> anansi_ID_sample_5_2   -1.226966212 -0.54920047
#> anansi_ID_sample_5_3   -1.072424501 -0.61108740
#> anansi_ID_sample_5_4    0.418516003 -1.61493436
#> anansi_ID_sample_6_1    1.416473124 -0.83330302
#> anansi_ID_sample_6_2   -0.192423751  1.01301284
#> anansi_ID_sample_6_3   -0.517516750 -1.65341623
#> anansi_ID_sample_6_4    0.418791957  0.39677330
#> anansi_ID_sample_7_1   -0.504971125  0.19605568
#> anansi_ID_sample_7_2    2.483726387  0.68791739
#> anansi_ID_sample_7_3    0.591350352  0.45110460
#> anansi_ID_sample_7_4   -1.049817584 -0.25092131
#> anansi_ID_sample_8_1    2.107440395  1.84061506
#> anansi_ID_sample_8_2    0.463411804 -2.07017799
#> anansi_ID_sample_8_3    1.864071931  2.31042863
#> anansi_ID_sample_8_4   -0.935702722 -0.13033969
#> anansi_ID_sample_9_1    2.368250631  1.08558881
#> anansi_ID_sample_9_2    0.552288536 -0.64707029
#> anansi_ID_sample_9_3    0.602568321  0.54073702
#> anansi_ID_sample_9_4   -0.628042145  0.15222802
#> anansi_ID_sample_10_1   0.094180829  0.34878330
#> anansi_ID_sample_10_2   0.257584617  1.36835363
#> anansi_ID_sample_10_3   0.368314603 -0.33594168
#> anansi_ID_sample_10_4  -0.074581204 -1.51260879
#> anansi_ID_sample_11_1  -0.515002451  0.19538164
#> anansi_ID_sample_11_2  -0.377963949 -1.66358698
#> anansi_ID_sample_11_3   0.117461881  0.27965076
#> anansi_ID_sample_11_4  -0.529431071  1.08009396
#> anansi_ID_sample_12_1   0.433582685  0.07356540
#> anansi_ID_sample_12_2  -0.791418858  1.35996771
#> anansi_ID_sample_12_3   1.938116055  1.18559450
#> anansi_ID_sample_12_4   0.184978630 -0.84674238
#> anansi_ID_sample_13_1  -0.677547109  1.13215387
#> anansi_ID_sample_13_2  -0.511219313 -1.42254276
#> anansi_ID_sample_13_3   1.570360027  0.32708435
#> anansi_ID_sample_13_4  -1.535114963 -0.91320074
#> anansi_ID_sample_14_1  -0.613917899 -1.87835910
#> anansi_ID_sample_14_2  -0.099601873 -1.38160048
#> anansi_ID_sample_14_3   0.054390122  1.24148261
#> anansi_ID_sample_14_4  -0.627601579 -0.32919594
#> anansi_ID_sample_15_1  -1.126570203  1.39464986
#> anansi_ID_sample_15_2   0.653453980 -0.33416496
#> anansi_ID_sample_15_3   1.698834092 -1.11246075
#> anansi_ID_sample_15_4  -0.241591640  0.64071473
#> anansi_ID_sample_16_1   0.909267302 -0.23181305
#> anansi_ID_sample_16_2   0.230977078  1.09464901
#> anansi_ID_sample_16_3  -0.550428864  0.10505241
#> anansi_ID_sample_16_4  -1.011731973 -2.42863359
#> anansi_ID_sample_17_1   0.488810749 -1.35990173
#> anansi_ID_sample_17_2  -0.072373896  0.17430555
#> anansi_ID_sample_17_3  -2.170218163  0.53314118
#> anansi_ID_sample_17_4  -0.080257435 -2.22282698
#> anansi_ID_sample_18_1   0.508082887  1.61733157
#> anansi_ID_sample_18_2   0.634677149 -0.41665420
#> anansi_ID_sample_18_3  -0.853328974  1.00952301
#> anansi_ID_sample_18_4  -0.779954949  0.23199690
#> anansi_ID_sample_19_1  -0.616191576 -0.46595652
#> anansi_ID_sample_19_2  -0.019196400 -0.16393813
#> anansi_ID_sample_19_3  -0.110991515  1.50745787
#> anansi_ID_sample_19_4  -0.273706160  0.22017918
#> anansi_ID_sample_20_1   0.983979281  0.39858082
#> anansi_ID_sample_20_2   1.029085968 -0.44897306
#> anansi_ID_sample_20_3   1.602425990 -0.06358934
#> anansi_ID_sample_20_4   0.808608575  2.00231226
#> anansi_ID_sample_21_1  -0.106083123  0.29357041
#> anansi_ID_sample_21_2   1.534257420 -2.10956413
#> anansi_ID_sample_21_3  -0.641491893  0.18086215
#> anansi_ID_sample_21_4  -1.655930807  0.17224010
#> anansi_ID_sample_22_1   0.906839593 -0.09425566
#> anansi_ID_sample_22_2   0.171764491  0.01229496
#> anansi_ID_sample_22_3  -0.287224258  1.04724999
#> anansi_ID_sample_22_4  -1.069443207 -0.22995775
#> anansi_ID_sample_23_1   0.671449505 -0.74784801
#> anansi_ID_sample_23_2   1.946295477  0.05096671
#> anansi_ID_sample_23_3  -0.646513840 -1.59656459
#> anansi_ID_sample_23_4   0.279928434 -0.60602637
#> anansi_ID_sample_24_1  -1.071927642  0.32792698
#> anansi_ID_sample_24_2   0.432535903 -0.48799653
#> anansi_ID_sample_24_3  -1.229990188  0.53346653
#> anansi_ID_sample_24_4   0.217741140 -1.70675208
#> anansi_ID_sample_25_1  -0.021018181  1.11428521
#> anansi_ID_sample_25_2   0.290133168 -0.59356477
#> anansi_ID_sample_25_3  -1.301782280  1.07375677
#> anansi_ID_sample_25_4   0.004930926  0.40465544
#> anansi_ID_sample_26_1  -0.419197962  1.24660407
#> anansi_ID_sample_26_2  -0.258272124 -1.11524545
#> anansi_ID_sample_26_3   0.387013494  0.35407631
#> anansi_ID_sample_26_4  -0.514148130  0.40088126
#> anansi_ID_sample_27_1  -1.328172982  0.56616481
#> anansi_ID_sample_27_2  -0.533051253 -0.70611414
#> anansi_ID_sample_27_3  -0.296472713  0.23189206
#> anansi_ID_sample_27_4   0.233670826  0.70034619
#> anansi_ID_sample_28_1  -0.545436915  0.22092862
#> anansi_ID_sample_28_2  -0.454328382  0.94493135
#> anansi_ID_sample_28_3   1.006300835 -1.59317389
#> anansi_ID_sample_28_4   0.260808067  0.12626472
#> anansi_ID_sample_29_1  -1.116993390  0.07962046
#> anansi_ID_sample_29_2  -1.814573296  1.19316824
#> anansi_ID_sample_29_3  -0.179786139 -0.31193914
#> anansi_ID_sample_29_4   0.792354718  2.05624227
#> anansi_ID_sample_30_1   0.693294878 -0.12571184
#> anansi_ID_sample_30_2  -1.226966212 -0.54920047
#> anansi_ID_sample_30_3  -1.072424501 -0.61108740
#> anansi_ID_sample_30_4   0.418516003 -1.61493436
#> anansi_ID_sample_31_1   1.416473124 -0.83330302
#> anansi_ID_sample_31_2  -0.192423751  1.01301284
#> anansi_ID_sample_31_3  -0.517516750 -1.65341623
#> anansi_ID_sample_31_4   0.418791957  0.39677330
#> anansi_ID_sample_32_1  -0.504971125  0.19605568
#> anansi_ID_sample_32_2   2.483726387  0.68791739
#> anansi_ID_sample_32_3   0.591350352  0.45110460
#> anansi_ID_sample_32_4  -1.049817584 -0.25092131
#> anansi_ID_sample_33_1   2.107440395  1.84061506
#> anansi_ID_sample_33_2   0.463411804 -2.07017799
#> anansi_ID_sample_33_3   1.864071931  2.31042863
#> anansi_ID_sample_33_4  -0.935702722 -0.13033969
#> anansi_ID_sample_34_1   2.368250631  1.08558881
#> anansi_ID_sample_34_2   0.552288536 -0.64707029
#> anansi_ID_sample_34_3   0.602568321  0.54073702
#> anansi_ID_sample_34_4  -0.628042145  0.15222802
#> anansi_ID_sample_35_1   0.094180829  0.34878330
#> anansi_ID_sample_35_2   0.257584617  1.36835363
#> anansi_ID_sample_35_3   0.368314603 -0.33594168
#> anansi_ID_sample_35_4  -0.074581204 -1.51260879
#> anansi_ID_sample_36_1  -0.515002451  0.19538164
#> anansi_ID_sample_36_2  -0.377963949 -1.66358698
#> anansi_ID_sample_36_3   0.117461881  0.27965076
#> anansi_ID_sample_36_4  -0.529431071  1.08009396
#> anansi_ID_sample_37_1   0.433582685  0.07356540
#> anansi_ID_sample_37_2  -0.791418858  1.35996771
#> anansi_ID_sample_37_3   1.938116055  1.18559450
#> anansi_ID_sample_37_4   0.184978630 -0.84674238
#> anansi_ID_sample_38_1  -0.677547109  1.13215387
#> anansi_ID_sample_38_2  -0.511219313 -1.42254276
#> anansi_ID_sample_38_3   1.570360027  0.32708435
#> anansi_ID_sample_38_4  -1.535114963 -0.91320074
#> anansi_ID_sample_39_1  -0.613917899 -1.87835910
#> anansi_ID_sample_39_2  -0.099601873 -1.38160048
#> anansi_ID_sample_39_3   0.054390122  1.24148261
#> anansi_ID_sample_39_4  -0.627601579 -0.32919594
#> anansi_ID_sample_40_1  -1.126570203  1.39464986
#> anansi_ID_sample_40_2   0.653453980 -0.33416496
#> anansi_ID_sample_40_3   1.698834092 -1.11246075
#> anansi_ID_sample_40_4  -0.241591640  0.64071473
#> anansi_ID_sample_41_1   0.909267302 -0.23181305
#> anansi_ID_sample_41_2   0.230977078  1.09464901
#> anansi_ID_sample_41_3  -0.550428864  0.10505241
#> anansi_ID_sample_41_4  -1.011731973 -2.42863359
#> anansi_ID_sample_42_1   0.488810749 -1.35990173
#> anansi_ID_sample_42_2  -0.072373896  0.17430555
#> anansi_ID_sample_42_3  -2.170218163  0.53314118
#> anansi_ID_sample_42_4  -0.080257435 -2.22282698
#> anansi_ID_sample_43_1   0.508082887  1.61733157
#> anansi_ID_sample_43_2   0.634677149 -0.41665420
#> anansi_ID_sample_43_3  -0.853328974  1.00952301
#> anansi_ID_sample_43_4  -0.779954949  0.23199690
#> anansi_ID_sample_44_1  -0.616191576 -0.46595652
#> anansi_ID_sample_44_2  -0.019196400 -0.16393813
#> anansi_ID_sample_44_3  -0.110991515  1.50745787
#> anansi_ID_sample_44_4  -0.273706160  0.22017918
#> anansi_ID_sample_45_1   0.983979281  0.39858082
#> anansi_ID_sample_45_2   1.029085968 -0.44897306
#> anansi_ID_sample_45_3   1.602425990 -0.06358934
#> anansi_ID_sample_45_4   0.808608575  2.00231226
#> anansi_ID_sample_46_1  -0.106083123  0.29357041
#> anansi_ID_sample_46_2   1.534257420 -2.10956413
#> anansi_ID_sample_46_3  -0.641491893  0.18086215
#> anansi_ID_sample_46_4  -1.655930807  0.17224010
#> anansi_ID_sample_47_1   0.906839593 -0.09425566
#> anansi_ID_sample_47_2   0.171764491  0.01229496
#> anansi_ID_sample_47_3  -0.287224258  1.04724999
#> anansi_ID_sample_47_4  -1.069443207 -0.22995775
#> anansi_ID_sample_48_1   0.671449505 -0.74784801
#> anansi_ID_sample_48_2   1.946295477  0.05096671
#> anansi_ID_sample_48_3  -0.646513840 -1.59656459
#> anansi_ID_sample_48_4   0.279928434 -0.60602637
#> anansi_ID_sample_49_1  -1.071927642  0.32792698
#> anansi_ID_sample_49_2   0.432535903 -0.48799653
#> anansi_ID_sample_49_3  -1.229990188  0.53346653
#> anansi_ID_sample_49_4   0.217741140 -1.70675208
#> anansi_ID_sample_50_1  -0.021018181  1.11428521
#> anansi_ID_sample_50_2   0.290133168 -0.59356477
#> anansi_ID_sample_50_3  -1.301782280  1.07375677
#> anansi_ID_sample_50_4   0.004930926  0.40465544
#> anansi_ID_sample_51_1  -0.419197962  1.24660407
#> anansi_ID_sample_51_2  -0.258272124 -1.11524545
#> anansi_ID_sample_51_3   0.387013494  0.35407631
#> anansi_ID_sample_51_4  -0.514148130  0.40088126
#> anansi_ID_sample_52_1  -1.328172982  0.56616481
#> anansi_ID_sample_52_2  -0.533051253 -0.70611414
#> anansi_ID_sample_52_3  -0.296472713  0.23189206
#> anansi_ID_sample_52_4   0.233670826  0.70034619
#> anansi_ID_sample_53_1  -0.545436915  0.22092862
#> anansi_ID_sample_53_2  -0.454328382  0.94493135
#> anansi_ID_sample_53_3   1.006300835 -1.59317389
#> anansi_ID_sample_53_4   0.260808067  0.12626472
#> anansi_ID_sample_54_1  -1.116993390  0.07962046
#> anansi_ID_sample_54_2  -1.814573296  1.19316824
#> anansi_ID_sample_54_3  -0.179786139 -0.31193914
#> anansi_ID_sample_54_4   0.792354718  2.05624227
#> anansi_ID_sample_55_1   0.693294878 -0.12571184
#> anansi_ID_sample_55_2  -1.226966212 -0.54920047
#> anansi_ID_sample_55_3  -1.072424501 -0.61108740
#> anansi_ID_sample_55_4   0.418516003 -1.61493436
#> anansi_ID_sample_56_1   1.416473124 -0.83330302
#> anansi_ID_sample_56_2  -0.192423751  1.01301284
#> anansi_ID_sample_56_3  -0.517516750 -1.65341623
#> anansi_ID_sample_56_4   0.418791957  0.39677330
#> anansi_ID_sample_57_1  -0.504971125  0.19605568
#> anansi_ID_sample_57_2   2.483726387  0.68791739
#> anansi_ID_sample_57_3   0.591350352  0.45110460
#> anansi_ID_sample_57_4  -1.049817584 -0.25092131
#> anansi_ID_sample_58_1   2.107440395  1.84061506
#> anansi_ID_sample_58_2   0.463411804 -2.07017799
#> anansi_ID_sample_58_3   1.864071931  2.31042863
#> anansi_ID_sample_58_4  -0.935702722 -0.13033969
#> anansi_ID_sample_59_1   2.368250631  1.08558881
#> anansi_ID_sample_59_2   0.552288536 -0.64707029
#> anansi_ID_sample_59_3   0.602568321  0.54073702
#> anansi_ID_sample_59_4  -0.628042145  0.15222802
#> anansi_ID_sample_60_1   0.094180829  0.34878330
#> anansi_ID_sample_60_2   0.257584617  1.36835363
#> anansi_ID_sample_60_3   0.368314603 -0.33594168
#> anansi_ID_sample_60_4  -0.074581204 -1.51260879
#> anansi_ID_sample_61_1  -0.515002451  0.19538164
#> anansi_ID_sample_61_2  -0.377963949 -1.66358698
#> anansi_ID_sample_61_3   0.117461881  0.27965076
#> anansi_ID_sample_61_4  -0.529431071  1.08009396
#> anansi_ID_sample_62_1   0.433582685  0.07356540
#> anansi_ID_sample_62_2  -0.791418858  1.35996771
#> anansi_ID_sample_62_3   1.938116055  1.18559450
#> anansi_ID_sample_62_4   0.184978630 -0.84674238
#> anansi_ID_sample_63_1  -0.677547109  1.13215387
#> anansi_ID_sample_63_2  -0.511219313 -1.42254276
#> anansi_ID_sample_63_3   1.570360027  0.32708435
#> anansi_ID_sample_63_4  -1.535114963 -0.91320074
#> anansi_ID_sample_64_1  -0.613917899 -1.87835910
#> anansi_ID_sample_64_2  -0.099601873 -1.38160048
#> anansi_ID_sample_64_3   0.054390122  1.24148261
#> anansi_ID_sample_64_4  -0.627601579 -0.32919594
#> anansi_ID_sample_65_1  -1.126570203  1.39464986
#> anansi_ID_sample_65_2   0.653453980 -0.33416496
#> anansi_ID_sample_65_3   1.698834092 -1.11246075
#> anansi_ID_sample_65_4  -0.241591640  0.64071473
#> anansi_ID_sample_66_1   0.909267302 -0.23181305
#> anansi_ID_sample_66_2   0.230977078  1.09464901
#> anansi_ID_sample_66_3  -0.550428864  0.10505241
#> anansi_ID_sample_66_4  -1.011731973 -2.42863359
#> anansi_ID_sample_67_1   0.488810749 -1.35990173
#> anansi_ID_sample_67_2  -0.072373896  0.17430555
#> anansi_ID_sample_67_3  -2.170218163  0.53314118
#> anansi_ID_sample_67_4  -0.080257435 -2.22282698
#> anansi_ID_sample_68_1   0.508082887  1.61733157
#> anansi_ID_sample_68_2   0.634677149 -0.41665420
#> anansi_ID_sample_68_3  -0.853328974  1.00952301
#> anansi_ID_sample_68_4  -0.779954949  0.23199690
#> anansi_ID_sample_69_1  -0.616191576 -0.46595652
#> anansi_ID_sample_69_2  -0.019196400 -0.16393813
#> anansi_ID_sample_69_3  -0.110991515  1.50745787
#> anansi_ID_sample_69_4  -0.273706160  0.22017918
#> anansi_ID_sample_70_1   0.983979281  0.39858082
#> anansi_ID_sample_70_2   1.029085968 -0.44897306
#> anansi_ID_sample_70_3   1.602425990 -0.06358934
#> anansi_ID_sample_70_4   0.808608575  2.00231226
#> anansi_ID_sample_71_1  -0.106083123  0.29357041
#> anansi_ID_sample_71_2   1.534257420 -2.10956413
#> anansi_ID_sample_71_3  -0.641491893  0.18086215
#> anansi_ID_sample_71_4  -1.655930807  0.17224010
#> anansi_ID_sample_72_1   0.906839593 -0.09425566
#> anansi_ID_sample_72_2   0.171764491  0.01229496
#> anansi_ID_sample_72_3  -0.287224258  1.04724999
#> anansi_ID_sample_72_4  -1.069443207 -0.22995775
#> anansi_ID_sample_73_1   0.671449505 -0.74784801
#> anansi_ID_sample_73_2   1.946295477  0.05096671
#> anansi_ID_sample_73_3  -0.646513840 -1.59656459
#> anansi_ID_sample_73_4   0.279928434 -0.60602637
#> anansi_ID_sample_74_1  -1.071927642  0.32792698
#> anansi_ID_sample_74_2   0.432535903 -0.48799653
#> anansi_ID_sample_74_3  -1.229990188  0.53346653
#> anansi_ID_sample_74_4   0.217741140 -1.70675208
#> anansi_ID_sample_75_1  -0.021018181  1.11428521
#> anansi_ID_sample_75_2   0.290133168 -0.59356477
#> anansi_ID_sample_75_3  -1.301782280  1.07375677
#> anansi_ID_sample_75_4   0.004930926  0.40465544
#> anansi_ID_sample_76_1  -0.419197962  1.24660407
#> anansi_ID_sample_76_2  -0.258272124 -1.11524545
#> anansi_ID_sample_76_3   0.387013494  0.35407631
#> anansi_ID_sample_76_4  -0.514148130  0.40088126
#> anansi_ID_sample_77_1  -1.328172982  0.56616481
#> anansi_ID_sample_77_2  -0.533051253 -0.70611414
#> anansi_ID_sample_77_3  -0.296472713  0.23189206
#> anansi_ID_sample_77_4   0.233670826  0.70034619
#> anansi_ID_sample_78_1  -0.545436915  0.22092862
#> anansi_ID_sample_78_2  -0.454328382  0.94493135
#> anansi_ID_sample_78_3   1.006300835 -1.59317389
#> anansi_ID_sample_78_4   0.260808067  0.12626472
#> anansi_ID_sample_79_1  -1.116993390  0.07962046
#> anansi_ID_sample_79_2  -1.814573296  1.19316824
#> anansi_ID_sample_79_3  -0.179786139 -0.31193914
#> anansi_ID_sample_79_4   0.792354718  2.05624227
#> anansi_ID_sample_80_1   0.693294878 -0.12571184
#> anansi_ID_sample_80_2  -1.226966212 -0.54920047
#> anansi_ID_sample_80_3  -1.072424501 -0.61108740
#> anansi_ID_sample_80_4   0.418516003 -1.61493436
#> anansi_ID_sample_81_1   1.416473124 -0.83330302
#> anansi_ID_sample_81_2  -0.192423751  1.01301284
#> anansi_ID_sample_81_3  -0.517516750 -1.65341623
#> anansi_ID_sample_81_4   0.418791957  0.39677330
#> anansi_ID_sample_82_1  -0.504971125  0.19605568
#> anansi_ID_sample_82_2   2.483726387  0.68791739
#> anansi_ID_sample_82_3   0.591350352  0.45110460
#> anansi_ID_sample_82_4  -1.049817584 -0.25092131
#> anansi_ID_sample_83_1   2.107440395  1.84061506
#> anansi_ID_sample_83_2   0.463411804 -2.07017799
#> anansi_ID_sample_83_3   1.864071931  2.31042863
#> anansi_ID_sample_83_4  -0.935702722 -0.13033969
#> anansi_ID_sample_84_1   2.368250631  1.08558881
#> anansi_ID_sample_84_2   0.552288536 -0.64707029
#> anansi_ID_sample_84_3   0.602568321  0.54073702
#> anansi_ID_sample_84_4  -0.628042145  0.15222802
#> anansi_ID_sample_85_1   0.094180829  0.34878330
#> anansi_ID_sample_85_2   0.257584617  1.36835363
#> anansi_ID_sample_85_3   0.368314603 -0.33594168
#> anansi_ID_sample_85_4  -0.074581204 -1.51260879
#> anansi_ID_sample_86_1  -0.515002451  0.19538164
#> anansi_ID_sample_86_2  -0.377963949 -1.66358698
#> anansi_ID_sample_86_3   0.117461881  0.27965076
#> anansi_ID_sample_86_4  -0.529431071  1.08009396
#> anansi_ID_sample_87_1   0.433582685  0.07356540
#> anansi_ID_sample_87_2  -0.791418858  1.35996771
#> anansi_ID_sample_87_3   1.938116055  1.18559450
#> anansi_ID_sample_87_4   0.184978630 -0.84674238
#> anansi_ID_sample_88_1  -0.677547109  1.13215387
#> anansi_ID_sample_88_2  -0.511219313 -1.42254276
#> anansi_ID_sample_88_3   1.570360027  0.32708435
#> anansi_ID_sample_88_4  -1.535114963 -0.91320074
#> anansi_ID_sample_89_1  -0.613917899 -1.87835910
#> anansi_ID_sample_89_2  -0.099601873 -1.38160048
#> anansi_ID_sample_89_3   0.054390122  1.24148261
#> anansi_ID_sample_89_4  -0.627601579 -0.32919594
#> anansi_ID_sample_90_1  -1.126570203  1.39464986
#> anansi_ID_sample_90_2   0.653453980 -0.33416496
#> anansi_ID_sample_90_3   1.698834092 -1.11246075
#> anansi_ID_sample_90_4  -0.241591640  0.64071473
#> anansi_ID_sample_91_1   0.909267302 -0.23181305
#> anansi_ID_sample_91_2   0.230977078  1.09464901
#> anansi_ID_sample_91_3  -0.550428864  0.10505241
#> anansi_ID_sample_91_4  -1.011731973 -2.42863359
#> anansi_ID_sample_92_1   0.488810749 -1.35990173
#> anansi_ID_sample_92_2  -0.072373896  0.17430555
#> anansi_ID_sample_92_3  -2.170218163  0.53314118
#> anansi_ID_sample_92_4  -0.080257435 -2.22282698
#> anansi_ID_sample_93_1   0.508082887  1.61733157
#> anansi_ID_sample_93_2   0.634677149 -0.41665420
#> anansi_ID_sample_93_3  -0.853328974  1.00952301
#> anansi_ID_sample_93_4  -0.779954949  0.23199690
#> anansi_ID_sample_94_1  -0.616191576 -0.46595652
#> anansi_ID_sample_94_2  -0.019196400 -0.16393813
#> anansi_ID_sample_94_3  -0.110991515  1.50745787
#> anansi_ID_sample_94_4  -0.273706160  0.22017918
#> anansi_ID_sample_95_1   0.983979281  0.39858082
#> anansi_ID_sample_95_2   1.029085968 -0.44897306
#> anansi_ID_sample_95_3   1.602425990 -0.06358934
#> anansi_ID_sample_95_4   0.808608575  2.00231226
#> anansi_ID_sample_96_1  -0.106083123  0.29357041
#> anansi_ID_sample_96_2   1.534257420 -2.10956413
#> anansi_ID_sample_96_3  -0.641491893  0.18086215
#> anansi_ID_sample_96_4  -1.655930807  0.17224010
#> anansi_ID_sample_97_1   0.906839593 -0.09425566
#> anansi_ID_sample_97_2   0.171764491  0.01229496
#> anansi_ID_sample_97_3  -0.287224258  1.04724999
#> anansi_ID_sample_97_4  -1.069443207 -0.22995775
#> anansi_ID_sample_98_1   0.671449505 -0.74784801
#> anansi_ID_sample_98_2   1.946295477  0.05096671
#> anansi_ID_sample_98_3  -0.646513840 -1.59656459
#> anansi_ID_sample_98_4   0.279928434 -0.60602637
#> anansi_ID_sample_99_1  -1.071927642  0.32792698
#> anansi_ID_sample_99_2   0.432535903 -0.48799653
#> anansi_ID_sample_99_3  -1.229990188  0.53346653
#> anansi_ID_sample_99_4   0.217741140 -1.70675208
#> anansi_ID_sample_100_1 -0.021018181  1.11428521
#> anansi_ID_sample_100_2  0.290133168 -0.59356477
#> anansi_ID_sample_100_3 -1.301782280  1.07375677
#> anansi_ID_sample_100_4  0.004930926  0.40465544
#> anansi::AnansiWeb S7_object with 400 observations:
#>     tableY: Enzyme (8 features)
#>     tableX: Metabolite (9 features)
#> Use @ to access: tableX, tableY, dictionary, metadata.
randomMultiFactor()
#> An anansi::MultiFactor S7_object,
#>     6 feature types across 5 edge lists.
#> 
#>       a   b   c   d   e   f
#> a2b 100 100   .   .   .   .
#> b2c   . 100 100   .   .   .
#> c2d   .   . 100 100   .   .
#> d2e   .   .   . 100 100   .
#> e2f   .   .   .   . 100 100
#> 
#> Values represent unique feature names in that edge list.
#> 
#> Levels:
#> 
#> a : 100 Levels: a_001 a_002 ... a_100 
#> b : 100 Levels: b_001 b_002 ... b_100 
#> c : 100 Levels: c_001 c_002 ... c_100 
#> d : 100 Levels: d_001 d_002 ... d_100 
#> e : 100 Levels: e_001 e_002 ... e_100 
#> f : 100 Levels: f_001 f_002 ... f_100