Generate a random AnansiWeb or AnansiLinkMap

Generate a random AnansiWeb or AnansiLinkMap

Usage

randomLinkMap(n_types = 6, n_features = 100, sparseness = 0.5)

randomWeb(
  n_samples = 10,
  n_features_x = 8,
  n_features_y = 12,
  sparseness = 0.5,
  tableY = NULL,
  tableX = NULL,
  dictionary = NULL
)

Arguments

n_types

Numeric scalar, number of types of features to generate

n_features

Numeric scalar, number of features per type

sparseness

Numeric scalar, proportion: How rare are connections

tableY, tableX

A table containing features of interest. Rows should be samples and columns should be features. Y and X refer to the position of the features in a formula: Y ~ X.

dictionary

A binary adjacency matrix of class Matrix, or coercible to Matrix

See also

Examples

# Make a random AnansiWeb object
randomWeb()
#> An object of class "AnansiWeb"
#> Slot "tableY":
#>            y
#> sample_id             y_1         y_2        y_3        y_4        y_5
#>   sample_1   1.8978176925  0.03982290 -1.7703502 -0.3605273 -0.7833267
#>   sample_2  -0.3024147872  0.08777483 -0.1614564 -0.9832709  0.4958025
#>   sample_3  -0.7487111833  0.19854754 -1.2006421 -0.6104247 -0.9728958
#>   sample_4  -1.5536886901  0.26232104  0.2426829 -0.2751353  0.4662975
#>   sample_5   0.0002394107 -1.25795301  0.6479128  0.1807655 -0.1287320
#>   sample_6  -0.1057841227  2.01132460 -0.8678791  0.2947642 -0.7162692
#>   sample_7  -1.1227589106 -0.07561685 -1.0961340  0.7780725 -0.6037589
#>   sample_8  -1.4279757590 -2.10454402  0.4579197  0.4685231  0.2002443
#>   sample_9  -0.7510236301  0.47041801  1.0289860  0.5896438 -0.7347615
#>   sample_10  0.1205262417 -0.38511470  0.7120576 -0.9082802  0.4022652
#>            y
#> sample_id          y_6         y_7         y_8        y_9       y_10
#>   sample_1  -0.5898533 -0.08336776  0.08424866  0.9238316 -0.2282311
#>   sample_2  -0.7021837 -1.33158502  0.49244422  0.7412008 -0.9901054
#>   sample_3  -0.1286774  0.53409599  0.35951074  0.1528107 -2.1257330
#>   sample_4   1.2269884  0.94557562 -0.98333181  0.3269209 -1.2540799
#>   sample_5   0.3545139 -1.89282039  0.03876090 -0.4832641 -0.1295985
#>   sample_6  -0.1545555  0.88315408  0.06266169 -0.2737629  0.4374770
#>   sample_7   1.3669988 -0.45053480 -1.23798323  0.3676516 -0.7589542
#>   sample_8  -0.1760295  0.25180893 -0.44899987 -0.7861871  0.7565663
#>   sample_9  -0.5193548 -0.22138022 -0.90860843 -1.1847536 -0.6458624
#>   sample_10 -1.0847083 -0.85444988 -1.29185659  1.0852679 -1.0350755
#>            y
#> sample_id          y_11        y_12
#>   sample_1  -0.02871595 -1.38689776
#>   sample_2  -1.40976906  0.76569946
#>   sample_3   0.02828043 -0.66321503
#>   sample_4  -1.34412200 -2.19597593
#>   sample_5  -0.25135352 -2.44005031
#>   sample_6   2.62566855  0.03494413
#>   sample_7  -0.78906530 -0.40491716
#>   sample_8  -0.55248661  0.14117142
#>   sample_9   0.25730542 -0.15592885
#>   sample_10 -0.14067576  0.33552558
#> 
#> Slot "tableX":
#>            x
#> sample_id           x_1        x_2        x_3         x_4         x_5
#>   sample_1  -0.53778021  0.1451637  1.7706075  0.55038434 -0.37569347
#>   sample_2  -0.21104669 -1.0907214 -0.4604116 -2.77543134  2.47133394
#>   sample_3   0.65501831 -0.6413128  0.2465064  0.30769261 -0.02785757
#>   sample_4  -0.31155970  0.5947005 -0.9997493 -1.61087052  0.54166705
#>   sample_5  -0.35408896  2.4428499  0.2765164  1.32647663 -1.32986830
#>   sample_6  -1.05889619  0.2475594 -0.5515622  1.19428844  0.73940048
#>   sample_7  -0.92620123  1.0673434  1.1234639  1.06509272 -0.67420561
#>   sample_8   1.07285711 -0.4105769  1.1637757 -1.61247894  0.75992770
#>   sample_9  -0.03283362 -0.9088979  0.1803260  0.23717921  0.56905446
#>   sample_10  2.11450519 -0.4895329 -0.6525869  0.03739383 -0.15521312
#>            x
#> sample_id          x_6          x_7        x_8
#>   sample_1  -0.2000290 -0.617825577  1.1744065
#>   sample_2   1.7536625 -0.785040344 -0.6803192
#>   sample_3   0.4877593 -0.001056473  0.2582959
#>   sample_4  -0.5494974 -1.416252938  0.4159256
#>   sample_5   1.8743386  0.107403757 -2.0246515
#>   sample_6  -0.4549758 -0.965004733 -1.0179494
#>   sample_7   0.3976505 -1.449620022 -0.1569594
#>   sample_8  -0.4710842  0.309887949 -1.1550237
#>   sample_9  -0.7351499 -0.070508493  0.5324802
#>   sample_10 -1.3067200  0.253909649 -0.2214920
#> 
#> Slot "dictionary":
#> 12 x 8 sparse Matrix of class "ngCMatrix"
#>       x
#> y      x_1 x_2 x_3 x_4 x_5 x_6 x_7 x_8
#>   y_1    |   .   |   .   .   .   .   .
#>   y_2    |   .   .   .   .   |   .   |
#>   y_3    |   |   |   .   |   |   .   .
#>   y_4    |   |   |   |   |   |   |   .
#>   y_5    |   |   .   .   .   .   .   |
#>   y_6    |   .   .   |   |   |   .   |
#>   y_7    .   .   |   .   |   .   .   .
#>   y_8    |   |   .   .   |   .   |   .
#>   y_9    .   .   |   |   |   .   |   .
#>   y_10   .   .   |   .   |   |   |   .
#>   y_11   |   |   .   |   |   |   |   .
#>   y_12   .   .   .   .   |   |   .   |
#> 
#> Slot "metadata":
#> DataFrame with 10 rows and 4 columns
#>                cat_ab     cat_XYZ   num_norm  num_unif
#>           <character> <character>  <numeric> <numeric>
#> sample_1            b           Z -0.8238259  0.195202
#> sample_2            b           X -1.0088838  0.983908
#> sample_3            b           Z  0.2155414  0.592911
#> sample_4            a           Y -0.6412864  0.739473
#> sample_5            a           Y  0.7379070  0.838061
#> sample_6            b           X -0.1991185  0.515351
#> sample_7            a           Y  0.7120941  0.696570
#> sample_8            b           X  1.5323565  0.634533
#> sample_9            a           X -1.2731041  0.655371
#> sample_10           a           X -0.0446331  0.310395
#> 
randomLinkMap()
#> An object of class AnansiLinkMap:
#> $a2b
#>   a   b 
#> 100 100 
#> 
#> $b2c
#>   b   c 
#> 100 100 
#> 
#> $c2d
#>   c   d 
#> 100 100 
#> 
#> $d2e
#>   d   e 
#> 100 100 
#> 
#> $e2f
#>   e   f 
#> 100 100 
#>